Health state transitions for continuous time state transition models

Summarize health state transitions in a continuous time state transition model with parameters that were estimated using a multi-state model.

CtstmTrans

Format

R6Class object.

Fields

data

Data used to predict health state transitions by strategy and patient. Must be an object of class input_data.

params

An object of class params_surv_list.

trans_mat

A transition matrix describing the states and transitions in a multi-state model in the format from the mstate package. See the documentation for the argument "trans" in msprep.

Methods

new(data, params)

Constructor for the class.

hazard(t)

Predict the hazard functions for each health state transition.

t: A numeric vector of times.

cumhazard(t)

Predict the cumulative hazard functions for each health state transition.

t: A numeric vector of times.

check()

Input validation for class. Checks that fields are the correct type.

Examples

# Simulation data
dt_strategies <- data.table(strategy_id = c(1, 2, 3))
dt_patients <- data.table(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))

# Multi-state model with transition specific models
fits <- vector(length = 3, mode = "list")
dat <- data.table(bosms3)
for (i in 1:length(fits)){
  fits[[i]] <- flexsurvreg(Surv(years, status) ~ 1, data = dat[trans == i],
                           dist = "exp")
}
fits <- flexsurvreg_list(fits)
tmat <- rbind(c(NA, 1, 2),
              c(NA, NA, 3),
              c(NA, NA, NA))

# Simulation model
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients)
fits_data <- expand_hesim_data(hesim_dat)
transmod <- create_CtstmTrans(fits, data = fits_data, trans_mat = tmat,
                              point_estimate = TRUE)
head(transmod$hazard(c(1, 2, 3)))

#>    trans sample strategy_id patient_id t    hazard
#> 1:     1      1           0          0 1 0.1898667
#> 2:     1      1           0          0 2 0.1898667
#> 3:     1      1           0          0 3 0.1898667
#> 4:     2      1           0          0 1 0.0534576
#> 5:     2      1           0          0 2 0.0534576
#> 6:     2      1           0          0 3 0.0534576

head(transmod$cumhazard(c(1, 2, 3)))

#>    trans sample strategy_id patient_id t cumhazard
#> 1:     1      1           0          0 1 0.1898667
#> 2:     1      1           0          0 2 0.3797333
#> 3:     1      1           0          0 3 0.5696000
#> 4:     2      1           0          0 1 0.0534576
#> 5:     2      1           0          0 2 0.1069152
#> 6:     2      1           0          0 3 0.1603728