Data tables for health economic simulation modeling

An R6Class containing the tables required for health-economic simulation modeling. Each table must either be an R data.frame or data.table.

DataTables

Format

R6Class object.

Fields

strategies

A table of the treatment strategies. Must contain the column strategy_id, denoting a unique strategy. Other columns denote characteristics of a treatment strategy.

patients

A table of patient observations. Must contain the column patient_id denoting a unique patient. The number of rows should be equal to the number of patients in the model. Other columns are variables describing the characteristics of a patient.

lines

A table of treatment lines used for each treatment strategy. Must contain the columns strategy_id, denoting a treatment strategy; line, denoting a treatment line; and treatment_id denoting the treatment used for a given strategy and line.

states

A table of health states. Must contain the column state_id, which denotes a unique health state. The number of rows should be equal to the number of health states in the model. Other columns can denote characteristics of a health state.

transitions

A table of health state transitions. Must contain the column transition_id, which denotes a unique transition; from, which denotes the starting health state; and to which denotes the state that will be transitioned to.

Methods

new(strategies, patients, states = NULL, transitions = NULL)

Constructor for the class.

validate()

Validate class fields.

expand(by = c("strategies", "patients"), ...)

Create a dataset where each row is a unique combination of the rows contained in the tables specified by character vector by. A wrapper around the merge function. The dataset will be sorted by prioritizing the id variables as follows: strategy_id, transition_id, patient_id, line, and state_id.

by: Data tables to expand by.

... Arguments to pass to merge.

Examples

dt.strategies <- data.table(strategy_id = c(1, 2))
dt.patients <- data.table(patient_id = seq(1, 3), age = c(65, 50, 75),
                          gender = c("Female", "Female", "Male"))
dt.lines <- lines_dt(list(c(1, 2, 5), c(1, 2)))
dt.states <- data.frame(state_id =  seq(1, 3),
                        state_var = c(2, 1, 9))
data.tables <- DataTables$new(strategies = dt.strategies,
                              patients = dt.patients,
                              states = dt.states,
                              lines = dt.lines)

#> Error in eval(expr, envir, enclos): object 'DataTables' not found

print(data.tables)

#> <DataTables>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     expand: function (by = c("strategies", "patients"), ...) 
#>     initialize: function (strategies, patients, lines = NULL, states = NULL, 
#>     lines: NULL
#>     patients: data.table, data.frame
#>     states: data.frame
#>     strategies: data.table, data.frame
#>     transitions: NULL
#>     validate: function () 

data.tables$expand(c("strategies", "patients"))

#>    strategy_id patient_id age female
#> 1:           1          1  45      0
#> 2:           1          2  50      0
#> 3:           1          3  60      1
#> 4:           2          1  45      0
#> 5:           2          2  50      0
#> 6:           2          3  60      1
#> 7:           3          1  45      0
#> 8:           3          2  50      0
#> 9:           3          3  60      1

data.tables$validate()

#> [1] "Validation successful"