Disease model based on overall survival

An R6 class for a modeling disease progression using survival analysis.

DisModSurv

Format

R6Class object.

Fields

dist_name

Name of the probability distribution used to model survival.

coefs

An object of class "surv_pars" representing the posterior distribution of coefficients from a survival model. For example, the output of rsample.

X

Matrix of covariates used to predict survival using the coefficients from surv_coefs.

time_length

Length of time units in years. Default is 1, which implies that time is measured in years. If time were measured in days, then time_length would be set to 1/365.

Members

new()

Create a new object of class Survival.

set_inputs(...)

Set the inputs for the class.

quantiles(q)

Generate the quantiles q of the survival distribution. q is a numeric vector.

surv(t)

Generate a survival curve give a numeric vector of times t.

cumhazard(t)

Generate cumulative hazards give a numeric vector of times t.

hazard(t)

Generate hazards give a numeric vector of times t.

rmst(t, r = 0)

Calculate the (discounted) restricted mean survival time. Calculated for a vector of times t given a discount rate r. Survival times are not discounted by default.