N-state partitioned survival model

Simulate otcomes from an N-state partitioned survival model.

PartSurv

Format

R6Class object.

Fields

survival_models

The survival models used to predict survival curves. Must be an object of class PartSurvCurves.

utility_model

The model used to predict utility by health state. Must be an object of class PartSurvStateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class PartSurvStateVals, where each element of the list represents a different type of costs.

n_states

Number of states in the partitioned survival model.

t_

A numeric vector of times at which survival curves were predicted. Determined by the argument t in sim_curves.

survival_

Survival curves generated using sim_curves.

stateprobs_

Health state probabilities as a function of time generated using sim_stateprobs.

costs_

Total (discounted) costs by type generated using sim_costs.

qalys_

Total (discounted) quality-adjusted life-years (QALYs) generated using sim_qalys.

Methods

new(survival_models, utility_model = NULL, cost_models = NULL)

Constructor for the class. Note that the number of health states, n_states, is set equal to the number of survival models plus one.

sim_survival(t)

Simulate survival curves as a function of time. Equivalent to the member function survival in PartSurvCurves.

t: A numeric vector of times. The first element must be 0.

sim_stateprobs()

Simulate the probability of being in each of N health states using the survival curves generated from sim_curves.

sim_qalys(dr = .03)

Simulate (discounted) QALYs over the times selected in t associated with each health state based on the state probabilities calculated using sim_stateprobs. See "Details".

dr: Discount rate to apply to QALYs. May be a vector in which case QALYs are calculated for each element in dr.

sim_costs(dr = .03)

Simulate (discounted) costs for each cost type over the times selected in t associated with each health state based on the state probabilities calculated using sim_stateprobs. See "Details".

dr: Discount rate to apply to costs. May be a vector in which case costs for each cost type are calculated for each element in dr.

check()

Input validation for class. Checks that fields are the correct type.

Details

Discounted costs and QALYs are calculated by integrating the "weighted" probability of being in each state. Weights are a function of the discount factor and the state value predicted using either the cost or QALY model. Mathematically, discounted costs and QALYs in health state \(s\) are calculated as, $$\int_0^T w_h e^{-rt} P_h(t) dt $$,

where for health state \(h\) and time t, \(w_h\) is the predicted cost or QALY weight, \(r\) is the discount rate, and \(P_h(t)\) is the probability of being in a given health state. The integral is calculated numerically using the composite trapezoid rule from the points in t_.

Examples

library("flexsurv")

# Simulation data
dt.strategies <- data.frame(strategy_id = c(1, 2, 3))
dt.patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))
dt.states <- data.frame(state_id =  seq(1, 3),
                           state_name = paste0("state", seq(1, 3)))
hesim.dat <- hesim_data(strategies = dt.strategies,
                              patients = dt.patients,
                              states = dt.states)
N <- 3

# Survival models
surv.data <- part_surv4_simdata$survival
fit1 <- flexsurv::flexsurvreg(Surv(endpoint1_time, endpoint1_status) ~ age,
                              data = surv.data, dist = "exp")
fit2 <- flexsurv::flexsurvreg(Surv(endpoint2_time, endpoint2_status) ~ age,
                              data = surv.data, dist = "exp")
fit3 <- flexsurv::flexsurvreg(Surv(endpoint3_time, endpoint3_status) ~ age,
                              data = surv.data, dist = "exp")
fits <- partsurvfit(flexsurvreg_list(fit1, fit2, fit3), data = surv.data)

curves.data <- expand_hesim_data(hesim.dat, by = c("strategies", "patients"))
part.surv.curves <- form_PartSurvCurves(fits, data = curves.data, n = N)

# Cost models
expanded.data <- expand_hesim_data(hesim.dat, by = c("strategies", "patients", "states"))
fit.costs.medical <- stats::lm(costs ~ female + state_name, data = part_surv4_simdata$costs$medical)
part.surv.costs.medical <- form_PartSurvStateVals(fit.costs.medical, data = expanded.data, n = N)

# Utility model
part.surv.utility.data <- form_input_data(formula_list(~1), expanded.data)
part.surv.utility <- PartSurvStateVals$new(data = part.surv.utility.data,
                                           params = params_lm(coef = runif(N, .6, .8)))

# Partitioned survival decision model
part.surv <- PartSurv$new(survival_models = part.surv.curves,
                          utility_model = part.surv.utility,
                          cost_models = list(medical = part.surv.costs.medical))
times <- seq(0, 5, .05)
part.surv$sim_survival(t = seq(0, 5, .05))
part.surv$sim_stateprobs()
part.surv$sim_costs(dr = .03)
head(part.surv$costs_)

#>    state_id sample strategy_id patient_id   dr    type    costs
#> 1:        1      1           1          1 0.03 medical 16785.56
#> 2:        1      1           1          2 0.03 medical 16006.68
#> 3:        1      1           1          3 0.03 medical 18705.41
#> 4:        1      1           2          1 0.03 medical 16785.56
#> 5:        1      1           2          2 0.03 medical 16006.68
#> 6:        1      1           2          3 0.03 medical 18705.41