Partitioned survival curves

Summarize n-1 survival curves for an n state partitioned survival model.

PartSurvCurves

Format

R6Class object.

Fields

data

Data used to predict state values by strategy and patient. Must be an object of class input_data where each row of a matrix in "X" is a unique strategy and patient. All matrices in "X" must be sorted by strategy and patient.

params

An object of class params_surv_list.

Methods

new(data, params)

Constructor for the class.

dataData used to summarize survival curves by strategy and patient. Must be an object of class input_data where each row in "X" is a strategy and patient pair.

paramsAn object of class partsurvfit, which contains the parameters for predicting survival.

hazard(t)

Predict the hazard function as a function of time.

t: A numeric vector of times.

cumhazard(t)

Predict the cumulative hazard function as a function of time.

t: A numeric vector of times.

survival(t)

Predict the survival function as a function of time.

t: A numeric vector of times.

rmst(t, dr = 0)

Predict (discounted) restricted mean survival time.

t: A numeric vector of times.

dr: Discount rate.

quantile(p)

Predict quantiles of the survival distributions.

p: A numeric vector of probabilities for calculating quantiles.

check()

Input validation for class; checks that fields are the correct type.

Examples

library("flexsurv")

# Simulation data
dt.strategies <- data.frame(strategy_id = c(1, 2, 3))
dt.patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))
hesim.dat <- hesim_data(strategies = dt.strategies,
                              patients = dt.patients)

# Fit survival models
surv.data <- part_surv4_simdata$survival
fit1 <- flexsurv::flexsurvreg(Surv(endpoint1_time, endpoint1_status) ~ age,
                              data = surv.data, dist = "exp")
fit2 <- flexsurv::flexsurvreg(Surv(endpoint2_time, endpoint2_status) ~ age,
                              data = surv.data, dist = "exp")
fit3 <- flexsurv::flexsurvreg(Surv(endpoint3_time, endpoint3_status) ~ age,
                              data = surv.data, dist = "exp")
fits <- partsurvfit(flexsurvreg_list(fit1, fit2, fit3), data = surv.data)

# Form PartSurvCurves
curves.data <- expand_hesim_data(hesim.dat, by = c("strategies", "patients"))
part.surv.curves <- form_PartSurvCurves(fits, data = curves.data, n = 3)

# Summarize survival curves
quantiles <- part.surv.curves$quantile(p = c(.25, .5, .75))
head(quantiles)

#>    curve sample strategy_id patient_id    p  quantile
#> 1:     1      1           1          1 0.25 0.1811498
#> 2:     1      1           1          1 0.50 0.4364661
#> 3:     1      1           1          1 0.75 0.8729322
#> 4:     1      1           1          2 0.25 0.1716965
#> 5:     1      1           1          2 0.50 0.4136891
#> 6:     1      1           1          2 0.75 0.8273782

survival <- part.surv.curves$survival(t = seq(0, 3, by = .1))
head(survival)

#>    curve sample strategy_id patient_id   t  survival
#> 1:     1      1           1          1 0.0 1.0000000
#> 2:     1      1           1          1 0.1 0.8531593
#> 3:     1      1           1          1 0.2 0.7278809
#> 4:     1      1           1          1 0.3 0.6209984
#> 5:     1      1           1          1 0.4 0.5298106
#> 6:     1      1           1          1 0.5 0.4520128

rmst <- part.surv.curves$rmst(t = c(2, 5))
head(rmst)

#>    curve sample strategy_id patient_id t      rmst
#> 1:     1      1           1          1 2 0.6034012
#> 2:     1      1           1          1 5 0.6294633
#> 3:     1      1           1          2 2 0.5759100
#> 4:     1      1           1          2 5 0.5966900
#> 5:     1      1           1          3 2 0.5233005
#> 6:     1      1           1          3 5 0.5361139