Survival models joined at different times

Predict or simulate values with survival model joined at different times.

StatModJoinedSurvs

Format

R6Class object.

Fields

input_data

An object of class InputData.

params

A list of object of class ParamsSurv.

times

A vector of times at which to join the survival parameters specified in params. The length of times should be equal to the length of params minus 1. Each element is the time to begin using new survival parameters; the first element in the params list is assumed to begin at time 0.

predict_type

The type of predicted value to be used. Current option is to simulate random values from the distribution specified in params ("random").

Members

new(input_data, params, times, predict_type)

Create a new object of class StatModSurvJoin. Models is a list of StatModSurv models and times

set_inputs(...)

Set the inputs for the class.

predict(tidy = FALSE)

Calculate mean survival time for each observation in input_data and each sample of parameters contained in params. If tidy = FALSE, then a matrix of predicted values is return; otherwise, a tidy data.table is returned.

simulate

Simulate survival times for each observation in input_data and each sample of parameters contained in params.

quantiles(q)

Generate the quantiles q of the survival distribution. q is a numeric vector.

surv(t)

Generate a survival curve give a numeric vector of times t.

cumhazard(t)

Generate cumulative hazards give a numeric vector of times t.

hazard(t)

Generate hazards give a numeric vector of times t.

rmst(t, dr = 0)

Calculate the (discounted) restricted mean survival time. Calculated for a vector of times t given a discount rate dr. Survival times are not discounted by default.