Form StateVals object

create_StateVals is a generic function for forming an object of class StateVals from a fitted statistical model.

create_StateVals(object, data, n = 1000, point_estimate = FALSE)

Arguments

object
A fitted statistical model object of the appropriate class. Supports lm.

data
An object of class "expanded_hesim_data" returned by expand_hesim_data. Must be expanded by the data tables "strategies", "patients", and "states".

n
Number of random observations of the parameters to draw.

point_estimate
If TRUE, then the point estimates are returned and and no samples are drawn.

...
Further arguments passed to or from other methods. Currently unused.

Value

Returns an R6Class object of class StateVals.

See also

StateVals