Create transition probabilities from array

Create a transition probability parameter object of class tparams_transprobs from a 4D array of matrices (i.e., a 6D array). The dimensions of the array should be as follows: 1st (sample), 2nd (strategy_id), 3rd (patient_id), 4th (time_id), 5th (rows of transition matrix), and 6th (columns of transition matrix). In other words, an index of [s, k, i, t] represents the transition matrix for the sth sample, kth treatment strategy, ith patient, and tth time interval.

# S3 method for array
create_tparams_transprobs(object, time_start = NULL)

Arguments

object
An object of class "array".

time_start
The starting time of each interval indexed by the 4th dimension of the array. This argument is not required if there is only 1 time interval as time_start will equal 0 in this case. Used to construct time_intervals with create_time_intervals.

Value

An object of class tparams_transprobs.

See also