Example data for a 3-state continuous time state transition model

A collection of example datasets containing health state transition, costs, and utility data for a 3-state continuous time state transition model.

ctstm3_exdata

Format

A list containing the following elements:

transitions A data frame containing the times at which patient transitions between health states based on the dataset prothr from the mstate package.

costs A list of data frames. The first data frame contains summary medical cost estimates and the second data frame contains drug cost data.

utility A data frame of summary utility estimates.

Transitions data

The data frame has the following columns:

strategy_id

Treatment strategy identification number.

patient_id

Patient identification number.

age

Patient age (in years).

female

1 if a patient is female; 0 if male.

from

Starting state.

to

Receiving state.

trans

Transition number.

Tstart

Starting time.

Tstop

Transition time.

years

Elapsed years between Tstart and Tstop.

status

Status variable; 1=transition, 0=censored.

Cost data

The cost list contains two data frames. The first data frame contains summary data on medical costs by health state, and contains the following columns:

state_id

The health state identification number.

mean

Mean costs.

se

Standard error of medical costs.

The second data frame contains data on the drug costs associated with each treatment strategy.

strategy_id

The treatment strategy identification number.

costs

Annualized drug costs.

Utility data

The data frame has the following columns:

state_id

The health state identification number.

mean

Mean costs.

se

Standard error.