Custom CEA summary table

Custom table summarizing outcomes from probabilistic sensitivity analysis.

custom_table(x, strategy, grp, custom_vars, FUN = NULL)

Arguments

x
Matrix containing information on outcome variables for each simulation and strategy.

strategy
Name of column denoting treatment strategy.

grp
Name of column denoting subgroup

custom_vars
Name of custom variables to calculate summary statistic for.

FUN
summary statistic function.

Value

A data.table of summary statistics for each variable specified in custom_vars. By default, returns the mean, 2.5%, and 97.5% quantile of each variable. Different summary statistics can be calculated using FUN. This function is used in icea and icea_pw to create the custom.table output.

Examples

# simulation output
nsims <- 100
sim <- data.frame(sim = rep(seq(nsims), 4),
             c = c(rlnorm(nsims, 5, .1), rlnorm(nsims, 5, .1),
                    rlnorm(nsims, 11, .1), rlnorm(nsims, 11, .1)),
             e = c(rnorm(nsims, 8, .2), rnorm(nsims, 8.5, .1),
                   rnorm(nsims, 11, .6), rnorm(nsims, 11.5, .6)),
             strategy = rep(paste0("Strategy ", seq(1, 2)),
                           each = nsims * 2),
             grp = rep(rep(c("Group 1", "Group 2"),
                           each = nsims), 2)
)

# Custom summary table
custom.fun <- function(x) list(mean = mean(x), median = median(x),
                             quantile(x, c(.025, .975)))
custom_table(sim, strategy = "strategy", grp = "grp",
           custom_vars = "e", FUN = custom.fun)

#>      strategy     grp    e.mean  e.median    e.2.5%   e.97.5%
#> 1: Strategy 1 Group 1  7.964688  7.946509  7.585361  8.351772
#> 2: Strategy 1 Group 2  8.494827  8.489791  8.277958  8.680028
#> 3: Strategy 2 Group 1 10.902032 10.900710  9.777534 12.246040
#> 4: Strategy 2 Group 2 11.533259 11.575093 10.361759 12.532697