Disease progression object

An object of class disprog returned from methods $sim_disease() in model classes. It contains simulated trajectories through a multi-state model.

Components

A disprog object inherits from data.table and contains the following columns:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

The patient ID.

from

The health state ID transitioned from.

to

The health state ID transitioned to.

final

An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time_start

The time at the start of the interval.

time_stop

The time at the end of the interval.

The object also contains size and absorbing attributes. The size attribute is a numeric vector with the elements n_samples, n_strategies, n_patients, and n_states denoting the number of samples, treatment strategies, patients, and health states The absorbing attribute is a numeric vector containing the absorbing health states; i.e., the health states that cannot be transitioned from. Operationally, an absorbing state is a row in a transition matrix (as in the trans_mat field of the IndivCtstmTrans class) with all NAs.

See also