Form InputData

form_InputData is a generic function for forming an R6Class object of class InputData; that is, the data containing the values of the input variables in the StatMod family of statistical models.

form_InputData(object, data_tables, ...)

# S3 method for formula
form_InputData(object, data_tables, ...)

# S3 method for formula_list
form_InputData(object, data_tables, ...)

# S3 method for lm
form_InputData(object, data_tables, ...)

# S3 method for flexsurvreg
form_InputData(object, data_tables, ...)

Arguments

object
An object of the appropriate class.

data_tables
An R6Class object of class DataTables.

...
Further arguments passed to model.matrix.

Value

An R6Class object of class InputData. The value of InputData$X depends on the class of its argument; it is either a matrix or a list of matrices used to predict the values of multiple parameters.

Examples

strategies <- list(c(1, 2, 3),
                   c(3, 5))
dt.strategy <- data.frame(strategy_dt(strategies))

#> Error in strategy_dt(strategies): could not find function "strategy_dt"

dt.strategy$strategy_var <- c(1, 2, 1, 4, 5)

#> Error in dt.strategy$strategy_var <- c(1, 2, 1, 4, 5): object 'dt.strategy' not found

dt.state <- data.frame(state_id =  seq(1, 3),
                       state_var = c(2, 1, 9))
dt.patient <- data.frame(patient_id = seq(1, 3), survival::lung[1:3, ])
db <- Database$new(strategy = dt.strategy,
                          state = dt.state,
                          patient = dt.patient)

#> Error in eval(expr, envir, enclos): object 'Database' not found

# form_InputData.formula
db$expand()

#> Error in eval(expr, envir, enclos): object 'db' not found

f <- formula(~ age)
input.dat <- form_InputData(f, db)

#> Error in form_InputData.formula(f, db): object 'db' not found

head(input.dat$X[[1]])

#> Error in head(input.dat$X[[1]]): object 'input.dat' not found

# form_InputData.formula_list
f.list <- formula_list(list(f1 = formula(~ age), f2 = formula(~ 1)))
input.dat <- form_InputData(f.list, db)

#> Error in form_InputData.formula_list(f.list, db): object 'db' not found

head(input.dat$X$f1)

#> Error in head(input.dat$X$f1): object 'input.dat' not found

head(input.dat$X$f2)

#> Error in head(input.dat$X$f2): object 'input.dat' not found

# form_InputData.lm
n <- 1000
x <- rbinom(n, 1, .53)
y <- 120 + 25 * x + rnorm(n, 0, 1)
patient.dt <- data.frame(x = x, y = y, id = seq(1, n))
db$set_inputs(patient = patient.dt)

#> Error in eval(expr, envir, enclos): object 'db' not found

db$expand()

#> Error in eval(expr, envir, enclos): object 'db' not found

fit <- stats::lm(y ~ x + strategy_var + state_var, data = db$data_)

#> Error in is.data.frame(data): object 'db' not found

input.dat <- form_InputData(fit, db)

#> Error in form_InputData(fit, db): object 'fit' not found

head(input.dat$X[[1]])

#> Error in head(input.dat$X[[1]]): object 'input.dat' not found

# form_InputData.flexsurvreg
patient.dt <- cbind(patient = seq(1, 3), flexsurv::bc[2:4, ])
db$set_inputs(patient = patient.dt)

#> Error in eval(expr, envir, enclos): object 'db' not found

db$expand()

#> Error in eval(expr, envir, enclos): object 'db' not found

fit <- flexsurv::flexsurvreg(Surv(recyrs, censrec) ~ group, data = bc,
                    anc = list(sigma = ~ group), dist = "gengamma")
input.dat <- form_InputData(fit, db)

#> Error in form_InputData.flexsurvreg(fit, db): object 'db' not found

head(input.dat$X$mu)

#> Error in head(input.dat$X$mu): object 'input.dat' not found

head(input.dat$X$sigma)

#> Error in head(input.dat$X$sigma): object 'input.dat' not found

head(input.dat$X$Q)

#> Error in head(input.dat$X$Q): object 'input.dat' not found