Parameters from a fitted model

form_params is a generic function for constructing parameters from a fitted statistical model. A sample from the posterior distribution of parameters is returned.

form_params(object, n = 1000, point_estimate = FALSE, ...)

# S3 method for lm
form_params(object, n = 1000, point_estimate = FALSE, ...)

# S3 method for flexsurvreg
form_params(object, n = 1000, point_estimate = FALSE,
  ...)

# S3 method for flexsurvreg_list
form_params(object, n = 1000,
  point_estimate = FALSE, ...)

# S3 method for partsurvfit
form_params(object, n = 1000, point_estimate = FALSE,
  bootstrap = TRUE, ...)

Arguments

object
A statistical model to randomly sample parameters from.

n
Number of random observations to draw.

point_estimate
If TRUE, then the point estimates are returned and and no samples are drawn.

...
Further arguments passed to or from other methods. Currently unused.

bootstrap
If bootstrap is FALSE or not specified, then n parameter sets are drawn by sampling from a multivariate normal distribution. If bootstrap is TRUE, then parameters are bootstrapped using bootstrap.

Value

An object prefixed by params_. Mapping between form_params and the classes of the returned objects are:

form_params.lm -> params_lm

form_params.flexsurvreg -> params_surv

form_params.flexsurvreg_list -> params_surv_list

form_params.partsurvfit -> params_surv_list

Examples

# form_params.lm
fit <- stats::lm(costs ~ female, data = part_surv4_simdata$costs$medical)
n <- 5
params.lm <- form_params(fit, n = n)
head(params.lm$coefs)

#>      (Intercept)     female
#> [1,]    31253.90  1598.3353
#> [2,]    33166.57 -1926.0922
#> [3,]    31444.63  1978.4934
#> [4,]    31226.01  -834.3395
#> [5,]    28898.51  3124.9976

head(params.lm$sigma)

#> [1] 10311.3 10311.3 10311.3 10311.3 10311.3

# form_params.flexsurvreg
library("flexsurv")
fit <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1,
                    data = ovarian, dist = "weibull")
n <- 5
params.surv <- form_params(fit, n = n)
print(params.surv$dist)

#> [1] "weibull.quiet"

head(params.surv$coefs)

#> $shape
#>           shape
#> [1,]  0.6204905
#> [2,] -0.1780169
#> [3,]  0.2572189
#> [4,]  0.1082616
#> [5,] -0.2551289
#> 
#> $scale
#>         scale
#> [1,] 6.759403
#> [2,] 6.921508
#> [3,] 6.868899
#> [4,] 7.272610
#> [5,] 7.916436
#>