List of survival distributions

List of additional distributions for parametric survival analysis that are not contained in flexsurv. Can be used to fit models with flexsurvreg. Same format as flexsurv.dists in flexsurv.

hesim_survdists

Format

A list with the following elements:

name

Name of the probability distribution.

pars

Vector of strings naming the parameters of the distribution. These must be the same names as the arguments of the density and probability functions.

location

Name of the location parameter.

transforms

List of R functions which transform the range of values taken by each parameter onto the real line. For example, c(log, log) for a distribution with two positive parameters.

inv.transforms

List of R functions defining the corresponding inverse transformations. Note these must be lists, even for single parameter distributions they should be supplied as, e.g. c(exp) or list(exp).

inits

A function of the observed survival times t (including right-censoring times, and using the halfway point for interval-censored times) which returns a vector of reasonable initial values for maximum likelihood estimation of each parameter. For example, function(t){ c(1, mean(t)) } will always initialize the first of two parameters at 1, and the second (a scale parameter, for instance) at the mean of t.