Individualized cost-effectiveness analysis

Conduct Bayesian cost-effectiveness analysis (e.g. summarize a probabilistic sensitivity analysis (PSA)) by subgroup.

icea() compute the probability that each treatment is most cost-effective, the expected value of perfect information, and the net monetary benefit for each treatment.

icea_pw() compares interventions to a comparator. Computed quantities include the incremental cost-effectiveness ratio, the incremental net monetary benefit, output for a cost-effectiveness plane, and output for a cost-effectiveness acceptability curve.

icea(x, k, sample = "sample", strategy = "strategy", grp = "grp",
  e = "e", c = "c")

icea_pw(x, k, comparator, sample = "sample", strategy = "strategy",
  grp = "grp", e = "e", c = "c")

Arguments

x
A data.frame or data.table containing simulation output with information on mean costs and effectiveness for each randomly sampled parameter set from a PSA. Each row should denote a randomly sampled parameter set and treatment strategy.

k
Vector of willingness to pay values

sample
Name of column denoting a randomly sampled parameter set. Default is "sample".

strategy
Name of column denoting treatment strategy. Default is "strategy".

grp
Name of column denoting subgroup. Default is "grp".

e
Name of column denoting clinical effectiveness. Default is "e".

c
Name of column denoting costs. Default is "c".

comparator
Name of the comparator strategy in x.

Value

icea returns a list containing four data.tables:

summary

A data.table of the mean, 2.5% quantile, and 97.5% quantile by strategy and group for clinical effectiveness and costs.

mce

The probability that each strategy is the most effective treatment for each group for the range of specified willingness to pay values.

evpi

The expected value of perfect information by group for the range of specified willingness to pay values.

nmb

The mean, 2.5% quantile, and 97.5% quantile of (monetary) net benefits for the range of specified willingness to pay values.

icea_pw also returns a list containing four data.tables:

summary

A data.table of the mean, 2.5% quantile, and 97.5% quantile by strategy and group for clinical effectiveness and costs.

delta

Incremental effectiveness and incremental cost for each simulated parameter set by strategy and group. Can be used to plot a cost-effectiveness plane.

ceac

Values needed to plot a cost-effectiveness acceptability curve by group. In other words, the probability that each strategy is more cost-effective than the comparator for the specified willingness to pay values.

inmb

The mean, 2.5% quantile, and 97.5% quantile of (monetary) incremental net benefits for the range of specified willingness to pay values.

Examples

# simulation output
n_samples <- 100
sim <- data.frame(sample = rep(seq(n_samples), 4),
                  c = c(rlnorm(n_samples, 5, .1), rlnorm(n_samples, 5, .1),
                        rlnorm(n_samples, 11, .1), rlnorm(n_samples, 11, .1)),
                  e = c(rnorm(n_samples, 8, .2), rnorm(n_samples, 8.5, .1),
                        rnorm(n_samples, 11, .6), rnorm(n_samples, 11.5, .6)),
                  strategy = rep(paste0("Strategy ", seq(1, 2)),
                                 each = n_samples * 2),
                  grp = rep(rep(c("Group 1", "Group 2"),
                            each = n_samples), 2)
)

# icea
icea_dt <- icea(sim, k = seq(0, 200000, 500), sample = "sample", strategy = "strategy",
 grp = "grp", e = "e", c = "c")
names(icea_dt)

#> [1] "summary" "mce"     "evpi"    "nmb"    

# the probability that each strategy is the most cost-effective 
# in each group with a willingness to pay of 20,000
library("data.table")

#> Warning: package ‘data.table’ was built under R version 3.4.4

icea_dt$mce[k == 20000]

#>        k   strategy     grp prob
#> 1: 20000 Strategy 1 Group 1 0.46
#> 2: 20000 Strategy 2 Group 1 0.54
#> 3: 20000 Strategy 1 Group 2 0.45
#> 4: 20000 Strategy 2 Group 2 0.55

# icea_pw
icea_pw_dt <-  icea_pw(sim,  k = seq(0, 200000, 500), comparator = "Strategy 1",
                       sample = "sample", strategy = "strategy", e = "e", c = "c")
names(icea_pw_dt)

#> [1] "summary" "delta"   "ceac"    "inmb"   

# cost-effectiveness acceptability curve
head(icea_pw_dt$ceac[k >= 20000])

#>        k   strategy     grp prob
#> 1: 20000 Strategy 2 Group 1 0.54
#> 2: 20000 Strategy 2 Group 2 0.55
#> 3: 20500 Strategy 2 Group 1 0.59
#> 4: 20500 Strategy 2 Group 2 0.59
#> 5: 21000 Strategy 2 Group 1 0.66
#> 6: 21000 Strategy 2 Group 2 0.60