Incremental cost-effectiveness ratio

Generate a tidy table of incremental cost-effectiveness ratios (ICERs) given output from cea_pw() with icer() and format for pretty printing with format.icer().

icer(x, prob = 0.95, k = 50000, labels = NULL, ...)

# S3 method for icer
format(
  x,
  digits_qalys = 2,
  digits_costs = 0,
  pivot_from = "strategy",
  drop_grp = TRUE,
  pretty_names = TRUE,
  ...
)

Arguments

x
An object of class cea_pw returned by cea_pw().

prob
A numeric scalar in the interval (0,1) giving the confidence interval. Default is 0.95 for a 95 percent interval.

k
Willingness to pay per quality-adjusted life-year.

labels
A list of named vectors containing the values and labels of variables. The elements of each vector are the values of a variable and the names are the labels. The names of the list are the names of the variables. See the output returned by get_labels() for an example.

...
Further arguments passed to and from methods. Currently unused.

digits_qalys
Number of digits to use to report QALYs.

digits_costs
Number of digits to use to report costs.

pivot_from
Character vector denoting a column or columns used to "widen" the data. Should either be "strategy", "grp", "outcome", or some combination of the three. There will be one column for each value of the variables in pivot_from. Default is to widen so there is a column for each treatment strategy. Set to NULL if you do not want to widen the table.

drop_grp
If TRUE, then the group column will be removed if there is only one subgroup; other it will be kept. If FALSE, then the grp column is never removed.

pretty_names
Logical. If TRUE, then the columns strategy, grp, outcome, and value are renamed (if they exist) to Strategy, Group, Outcome, and Value.

Value

icer() returns an object of class icer that is a tidy data.table with the following columns:

strategy

The treatment strategy.

grp

The subgroup.

outcome

The outcome metric.

estimate

The point estimate computed as the average across the PSA samples.

lower

The lower limit of the confidence interval.

upper

The upper limit of the confidence interval.

format.icer() formats the table according to the arguments passed.

Details

Note that icer() will report negative ICERs; however, format() will correctly note whether a treatment strategy is dominated by or dominates the reference treatment.

See also