Data tables for health economic simulation modeling

An R6Class containing the tables required for health-economic simulation modeling. Each table must either be an R data.frame or data.table.

form(fit)

Format

R6Class object.

Fields

strategy

A table of the treatment strategies. Must contain the column strategy_id, denoting a unique strategy. May also contain the column line, denoting the line of therapy associated with a given strategy, so that each observation in the table is a (strategy, line) pair. Other columns denote characteristics of a stategy or a (strategy, line) pair.

patient

A table of patient observations. Must contain the column patient_id denoting a unique patient. The number of rows should be equal to the number of patients in the model. Other columns are variables describing the characteristics of a patient.

state

A table of health states. Must contain the column state_id, which denotes a unique health state. The number of rows should be equal to the number of health states in the model. Other columns can denote characteristics of a health state.

transitions

A table of health state transitions. Must contain the column trans_id, which denotes a unique transition; from, which denotes the starting health state; and to which denotes the state that will be transitioned to.

data_

A dataset generated using expand().

by_

A character vector denoting the tables used to generate data_.

Members

new(strategy, patient, state = NULL, transitions = NULL)

Create a new object of class DataTables.

set_inputs(...)

Set the inputs for the class.

expand(tables = c("strategy", "patient"), ...)

Create a dataset where each row is a unique commbination of the rows contained in the tables specified by character vector tables. A wrapper around the merge function. Use ... to pass arguments to merge.

Examples

library("survival")
strategies <- list(c(1, 2, 3),
                  c(3, 5))
dt.strategy <- data.frame(strategy_dt(strategies))
dt.strategy$strategy_var <- c(1, 2, 1, 4, 5)
dt.patient <- data.frame(patient_id = seq(1, 3), survival::lung[1:3, ])
data.tables <- DataTables(strategy = dt.strategy,
                      patient = dt.patient)

#> Error in DataTables(strategy = dt.strategy, patient = dt.patient): could not find function "DataTables"

print(data.tables)

#> <input_data_DataTables>
#>   Public:
#>     by_: active binding
#>     clone: function (deep = FALSE) 
#>     data_: active binding
#>     expand: function (tables = c("strategy", "patient"), ...) 
#>     initialize: function (strategy, patient, lines = NULL, state = NULL, transitions = NULL) 
#>     lines: NULL
#>     patient: data.table, data.frame
#>     set_inputs: function (...) 
#>     state: data.frame
#>     strategy: data.table, data.frame
#>     transitions: NULL
#>   Private:
#>     .by_: strategy patient
#>     .data_: data.table, data.frame

data.tables$expand(c("strategy", "patient"))$data_

#>    strategy_id patient_id inst time status age sex ph.ecog ph.karno pat.karno
#> 1:           1          1    3  306      2  74   1       1       90       100
#> 2:           2          1    3  306      2  74   1       1       90       100
#> 3:           1          2    3  455      2  68   1       0       90        90
#> 4:           2          2    3  455      2  68   1       0       90        90
#> 5:           1          3    3 1010      1  56   1       0       90        90
#> 6:           2          3    3 1010      1  56   1       0       90        90
#>    meal.cal wt.loss
#> 1:     1175      NA
#> 2:     1175      NA
#> 3:     1225      15
#> 4:     1225      15
#> 5:       NA      15
#> 6:       NA      15