Join statistical models at specified times

join is a generic function for constructing classes, which are used to join statistical models at specified time points. The first model is used prior to the first time point, the second model is used prior to the second time point, and so on. There should be N + 1 models and N time points.

join(objects, times)

Arguments

objects
A list of model objects to join. Currently there are methods for formula_list, formula_list_2d, flexsurvreg_list, and flexsurvreg_list_2d.

times
Times to join models. If a list of models is specified in objects, then a vector of times. If a list of lists of models is specified in objects, then a list of vectors of times.

Value

Returns a list with two variables:

models

List of joined models.

times

Time points to join models.

The relationships between the generic functions and the returned classes are:

join.flexsurvreg_list -> "joined_flexsurvreg"

join.flexsurvreg_list_2d -> "joined_flexsurvreg_list"

Examples

fit.wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1,
                                 data = ovarian, dist = "weibull")
fit.exp <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1,
                             data = ovarian, dist = "exp")
flexsurvreg.list <- flexsurvreg_list(wei = fit.wei, exp = fit.exp)
joined.flexsurvreg <- join(flexsurvreg.list, times = 10)
class(joined.flexsurvreg)

#> [1] "joined_flexsurvreg"