#' Data for StatMod statistical models #' #' An R6Class of data containing the values of input variables #' used to predict or simulate outputs with the StatMod family of #' statistical models. #' @format R6Class object. #' @field X A list of matrices containing values of covariates used to predict or #' simulate outputs. The list will be of length 1 when the covariates are used #' to predict a single parameter or outcome; in models where covariates are used #' to predict multiple parameters (e.g., parametric survival models), the list #' will be of length > 1. #' @field strategy_id Vector of values denoting a treatment strategy. Length is equal #' to the number of rows in X. #' @field line Vector of values denoting a line of therapy. Length is equal to #' the number of rows in X. #' @field state_id Vector of values denoting a disease state. Length is equal to the #' number of rows in X. #' @field patient_id Vector of values denoting a patient. Length is equal to the #' number of rows in X. #' @field timefun An external pointer to a compiled C++ functor used to #' update time-dependent covariates in X. Must inherit from the #' hesim::TimeFunc class. #' @field sorted Are the rows of X sorted by strategy, #' line, state, and patient? #' @field n_strategies An integer denoting the of distinct treatment strategies. #' @field n_lines A data.frame containing the number of lines for each #' treatment strategy. Column 1 is strategy_id and column 2 is N, or, #' the number of lines. #' @field n_states An integer denoting number of distinct disease states. #' @field n_patients An integer denoting number of distinct patients. #' #' @section Members: #'
#'
new(X, state_id, strategy_id, line_id, patient_id, #' time_fun = NULL, n_strategies = NULL, #' n_lines = NULL, n_states = NULL, n_patients = NULL, sorted = FALSE)

Create a new object of class InputData.

#'
sort()

Sort rows of X by strategy_id, line, state_id, #' and patient_id.

#'
#' @export InputData <- R6::R6Class("InputData", public = list( X = NULL, strategy_id = NULL, line = NULL, state_id = NULL, patient_id = NULL, timefun = NULL, n_strategies = NULL, n_lines = NULL, n_states = NULL, n_patients = NULL, n_obs = NULL, sorted = NULL, initialize = function(X, strategy_id, line, state_id, patient_id, timefun = NULL, n_strategies = NULL, n_lines = NULL, n_states = NULL, n_patients = NULL, sorted = FALSE) # initialize self$X <- X self$state_id <- state_id self$strategy_id <- strategy_id self$line <- line self$patient_id <- patient_id self$timefun <- timefun self$n_strategies <- len_InputData(n_strategies, self$strategy_id) self$n_lines <- if (!is.null(n_lines)) n_lines else lines.unique <- unique(data.table(strategy_id = self$strategy_id, line = self$line)) as.data.frame(lines.unique[, .N, by = "strategy_id"]) self$n_states <- len_InputData(n_states, self$state_id) self$n_patients <- len_InputData(n_patients, self$patient_id) self$sorted <- sorted , sort = function() index <- order(self$strategy_id, self$line, self$state_id, self$patient_id) self$strategy_id <- self$strategy_id[index] self$line <- self$line[index] self$state_id <- self$state_id[index] self$patient_id <- self$patient_id[index] self$n_lines <- self$n_lines[order(self$n_lines$strategy_id), ] self$X <- sort_InputData_X(self$X, index) self$sorted <- TRUE invisible(self) ) )

#' Data for StatMod statistical models #' #' An R6Class of data containing the values of input variables #' used to predict or simulate outputs with the StatMod family of #' statistical models. #' @format R6Class object. #' @field X A list of matrices containing values of covariates used to predict or #' simulate outputs. The list will be of length 1 when the covariates are used #' to predict a single parameter or outcome; in models where covariates are used #' to predict multiple parameters (e.g., parametric survival models), the list #' will be of length > 1. #' @field strategy_id Vector of values denoting a treatment strategy. Length is equal #' to the number of rows in X. #' @field line Vector of values denoting a line of therapy. Length is equal to #' the number of rows in X. #' @field state_id Vector of values denoting a disease state. Length is equal to the #' number of rows in X. #' @field patient_id Vector of values denoting a patient. Length is equal to the #' number of rows in X. #' @field timefun An external pointer to a compiled C++ functor used to #' update time-dependent covariates in X. Must inherit from the #' hesim::TimeFunc class. #' @field sorted Are the rows of X sorted by strategy, #' line, state, and patient? #' @field n_strategies An integer denoting the of distinct treatment strategies. #' @field n_lines A data.frame containing the number of lines for each #' treatment strategy. Column 1 is strategy_id and column 2 is N, or, #' the number of lines. #' @field n_states An integer denoting number of distinct disease states. #' @field n_patients An integer denoting number of distinct patients. #' #' @section Members: #'

#'
new(X, state_id, strategy_id, line_id, patient_id, #' time_fun = NULL, n_strategies = NULL, #' n_lines = NULL, n_states = NULL, n_patients = NULL, sorted = FALSE)

Create a new object of class InputData.

#'
sort()

Sort rows of X by strategy_id, line, state_id, #' and patient_id.

#'

#' @export InputData <- R6::R6Class("InputData", public = list( X = NULL, strategy_id = NULL, line = NULL, state_id = NULL, patient_id = NULL, timefun = NULL, n_strategies = NULL, n_lines = NULL, n_states = NULL, n_patients = NULL, n_obs = NULL, sorted = NULL,

initialize = function(X, strategy_id, line, state_id, patient_id, timefun = NULL, n_strategies = NULL, n_lines = NULL, n_states = NULL, n_patients = NULL, sorted = FALSE) # initialize self$X <- X self$state_id <- state_id self$strategy_id <- strategy_id self$line <- line self$patient_id <- patient_id self$timefun <- timefun self$n_strategies <- len_InputData(n_strategies, self$strategy_id) self$n_lines <- if (!is.null(n_lines)) n_lines else lines.unique <- unique(data.table(strategy_id = self$strategy_id, line = self$line)) as.data.frame(lines.unique[, .N, by = "strategy_id"]) self$n_states <- len_InputData(n_states, self$state_id) self$n_patients <- len_InputData(n_patients, self$patient_id) self$sorted <- sorted ,

sort = function() index <- order(self$strategy_id, self$line, self$state_id, self$patient_id) self$strategy_id <- self$strategy_id[index] self$line <- self$line[index] self$state_id <- self$state_id[index] self$patient_id <- self$patient_id[index] self$n_lines <- self$n_lines[order(self$n_lines$strategy_id), ] self$X <- sort_InputData_X(self$X, index) self$sorted <- TRUE invisible(self) ) )

len_InputData(len = NULL, id)