Simulate state transitions in markov cohort model

This function simulates state transitions in a Markov cohort model. Allows for the simulation of multiple cohorts. The model is fully deterministic and is appropriate for estimating population averages in each cohort.

markov_cohort_trans(z0, ncycles, pmat, pmat_index = NULL, mortadj = FALSE,
  mortprob = NULL, mortprob_index = NULL)

Arguments

z0

Matrix of starting distribution. One row for each cohort and one column for each model state including abosrbing (i.e. death) states. If mortadj = TRUE, then z0 should have one column for each model state in pmat and an additional column for the state of death.

ncycles

Vector with the ith element indicating the number cycles to run the model for the ith row in z0.

pmat

Array of distinct transition probability matrices.

mortadj

Logical. If TRUE, then the a death state is added to the probability matrix and probabilities are adjusted accordingly. If FALSE, then probability matrix remains as is.

mortprob

Array of distinct mortality probability matrices. Each slice of array contains a matrix with rows denoting model cycle (or patient age) and columns representing model state. Each element in the matrix is the probability of mortality for a particular group. For example, a particular element might contain the probability of mortality for males (3rd dimension of array) at age 55 (1st dimension of array) in model state 1 (2nd dimension of array).

mortprob_index

Index vector denoting the mortality probability matrix to apply to each cohort. If NULL, the 3rd dimension of mortprob must equal the number of rows in z0.

p_index

Index vector denoting the probability matrix to apply to each cohort. If NULL, the 3rd dimension of pmat must equal number of rows in z0.

Value

Matrix with the number of simulated individuals in each model state for each cohort during each cycle.

Details

The code is written in c++ to minmize simulation time.