Method of moments for beta distribution

Compute the parameters shape1 and shape2 of the beta distribution using method of moments given the mean and standard deviation of the random variable of interest.

mom_beta(mean, sd)

Arguments

mean

Mean of the random variable.

sd

Standard deviation of the random variable.

Value

A list containing the parameters shape1 and shape2.

Details

If \(\mu\) is the mean and \(\sigma\) is the standard deviation of the random variable, then the method of moments estimates of the parameters shape1 = \(\alpha > 0\) and shape2 = \(\beta > 0\) are: $$\alpha = \mu \left(\frac{\mu(1-\mu)}{\sigma^2}-1 \right)$$ and $$\beta = (1 - \mu) \left(\frac{\mu(1-\mu)}{\sigma^2}-1 \right)$$

Examples

mom_beta(mean = .8, sd = .1)
#> $shape1
#> [1] 12
#> 
#> $shape2
#> [1] 3
#> 
# The function is vectorized.
mom_beta(mean = c(.6, .8), sd = c(.08, .1))
#> $shape1
#> [1] 21.9 12.0
#> 
#> $shape2
#> [1] 14.6  3.0
#>