Example data for a 3-state multinomial model

Example discrete time health state transitions data simulated using multinomial logistic regression. Costs and utility are also included to facilitate cost-effectiveness analysis.

multinom3_exdata

Format

A list containing the following elements:

transitions

A data frame containing patient transitions between health states at discrete time intervals (i.e., on a yearly basis).

costs

A list of data frames. The first data frame contains drug cost data and the second contains summary medical cost estimates.

utility

A data frame of summary utility estimates.

Transitions data

The data frame has the following columns:

patient_id

Patient identification number.

strategy_id

Treatment strategy identification number.

strategy_name

Treatment strategy name.

age

Patient age (in years).

age_cat

A factor variable with 3 age groups: (i) age less than 40, (ii) age at least 40 and less than 60, and (iii) age at least 60.

female

1 if a patient is female; 0 if male.

year

The year since the start of data collection with the first year equal to 1.

state_from

State making a transition from.

state_to

State making a transition to.

year_cat

Factor variable for year with 3 categories: (i) year 3 and below, (ii) year between 3 and 6, and (iii) year 7 and above.

Cost data

The cost list contains two data frames. The first data frame contains data on the drug costs associated with each treatment strategy.

strategy_id

The treatment strategy identification number.

strategy_name

The treatment strategy name.

costs

Annualized drug costs.

The second data frame contains summary data on medical costs by health state, and contains the following columns:

state_id

The health state identification number.

state_name

The name of the health state.

mean

Mean medical costs.

se

Standard error of medical costs.

Utility data

The data frame has the following columns:

state_id

The health state identification number.

state_name

The name of the health state.

mean

Mean utility

se

Standard error of utility.