Multi-state panel oncology data for 3-state model

The same dataset as onc3 converted into a panel data format in which health states are recorded at a finite series of times.

onc3p

Format

A data.table with the following columns:

state

The name of the health state (Stable, Progression, and Death).

strategy_name

Standard of care (SOC), new treatment 1 (New 1), or new treatment 2 (New 2).

female

1 if a patient is female; 0 if male.

age

Patient age (in years).

patient_id

Patient identification number.

time

Time that state was recorded.

strategy_id

Strategy identification number.

state_id

The health state identification number.

See also

Examples

head(onc3p)
#>          state strategy_name female      age patient_id      time strategy_id
#>         <char>        <fctr>  <int>    <num>      <int>     <num>       <int>
#> 1:      Stable         New 2      0 59.85813          1  0.000000           3
#> 2: Progression         New 2      0 59.85813          1  2.420226           3
#> 3:       Death         New 2      0 59.85813          1 14.620258           3
#> 4:      Stable         New 2      0 62.57282          2  0.000000           3
#> 5:      Stable         New 2      0 62.57282          2  7.497464           3
#> 6:      Stable           SOC      1 61.44379          3  0.000000           1
#>    state_id
#>      <fctr>
#> 1:        1
#> 2:        2
#> 3:        3
#> 4:        1
#> 5:        1
#> 6:        1