Parameters of joined lists of survival models

Create a list containing the parameters of multiple sets of survival models, each joined at specified time points. See joined for more details.

params_joined_surv_list(..., times)

Arguments

...
Objects of class params_surv_list, which can be named.

times
A list of sorted numeric vectors, with the length of each list element equal to the number of sets of models.

Value

An object of class params_joined_surv_list, which is a list containing two elements:

models

A list of params_surv_list, each containing params_surv objects to be joined.

times

Equivalent to the argument times.

Examples

library("flexsurv")
fit_exp <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "exp")
fit_wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "weibull")
fit_lnorm <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                  data = ovarian, dist = "lognormal")

params_exp <- create_params(fit_exp, n = 2)
params_wei <- create_params(fit_wei, n = 2)
params_lnorm <- create_params(fit_lnorm, n = 2)

params_list1 <- params_surv_list(params_exp, params_wei)
params_list2 <- params_surv_list(params_exp, params_lnorm)
params_joined <- params_joined_surv_list(model1 = params_list1,
                                         model2 = params_list2,
                                         times = list(3, 5))
print(params_joined)

#> $models
#> $models$model1
#> [[1]]
#> $coefs
#> $coefs$rate
#>      (Intercept)
#> [1,]   -7.223312
#> [2,]   -7.151825
#> 
#> 
#> $dist
#> [1] "exp"
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_surv"
#> 
#> [[2]]
#> $coefs
#> $coefs$shape
#>      (Intercept)
#> [1,]   0.0569819
#> [2,]   0.1693948
#> 
#> $coefs$scale
#>      (Intercept)
#> [1,]    7.360524
#> [2,]    6.479149
#> 
#> 
#> $dist
#> [1] "weibull.quiet"
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_surv"
#> 
#> attr(,"class")
#> [1] "params_surv_list"
#> 
#> $models$model2
#> [[1]]
#> $coefs
#> $coefs$rate
#>      (Intercept)
#> [1,]   -7.223312
#> [2,]   -7.151825
#> 
#> 
#> $dist
#> [1] "exp"
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_surv"
#> 
#> [[2]]
#> $coefs
#> $coefs$meanlog
#>      (Intercept)
#> [1,]    7.300886
#> [2,]    6.394735
#> 
#> $coefs$sdlog
#>      (Intercept)
#> [1,]  0.61132222
#> [2,]  0.05410243
#> 
#> 
#> $dist
#> [1] "lnorm"
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_surv"
#> 
#> attr(,"class")
#> [1] "params_surv_list"
#> 
#> 
#> $times
#> $times[[1]]
#> [1] 3
#> 
#> $times[[2]]
#> [1] 5
#> 
#> 
#> attr(,"class")
#> [1] "params_joined_surv_list"