Parameters of a linear model

Create a list containing the parameters of a fitted linear regression model.

params_lm(coefs, sigma = NULL)

Arguments

coefs
Matrix of samples from the posterior distribution of the regression coefficients.

sigma
A vector of samples of the standard error of the regression model.

Value

An object of class "params_lm", which is a list containing coefs, sigma, and n_samples. n_samples is equal to the number of rows in coefs.

Examples

library("MASS")
n <- 2
params <- params_lm(coefs = MASS::mvrnorm(n, mu = c(.5,.6),
                                            Sigma = matrix(c(.05, .01, .01, .05), nrow = 2)),
                      sigma <- rgamma(n, shape = .5, rate = 4))
print(params)

#> $coefs
#>           [,1]      [,2]
#> [1,] 0.7655770 0.4965964
#> [2,] 0.4585277 0.7672023
#> 
#> $sigma
#> [1] 0.66551976 0.07453363
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_lm"