Parameters for state values

Parameters that are used to simulate values assigned to health states with StateVals.

params_statevals(values, timefun = NULL)

Arguments

values
Typically a matrix where each column denotes a health state and each row denotes a random sample of the value assigned to that health state for the probabilistic sensitivitiy analysis. If health state values vary by strategy, then it can be an array of matrices where each matrix denotes the values associated with a different treatment strategy.

timefun
An R function that can be used to modify values as a function of time.

Value

An object of class "params_statevals", which is a list containing values and timefun.

Examples

# Cost estimates in 2 health states
gamma_params <- mom_gamma(c(5000, 7000), c(1000, 1200))
n <- 1000
vals <- matrix(rgamma(2 * n,
                      shape = gamma_params$shape,
                      scale = gamma_params$scale),
               nrow = n, ncol = 2, byrow = TRUE)
params_stvals <- params_statevals(values = vals)
print(params_stvals)

#> $values
#>             [,1]      [,2]
#>    [1,] 4456.966  7112.526
#>    [2,] 4677.471  7204.636
#>    [3,] 4068.667  8497.707
#>    [4,] 5312.908  7915.221
#>    [5,] 3683.385  5763.850
#>    [6,] 7214.563  5838.620
#>    [7,] 4995.061  8058.870
#>    [8,] 4159.543  6570.296
#>    [9,] 4250.528  7273.449
#>   [10,] 7547.323  7700.001
#>   [11,] 5938.400  6145.567
#>   [12,] 3543.228  5686.357
#>   [13,] 6064.064  6386.358
#>   [14,] 6645.716  7826.329
#>   [15,] 5317.974  4829.869
#>   [16,] 6509.226  6886.364
#>   [17,] 4020.967  7329.916
#>   [18,] 5242.873 11079.758
#>   [19,] 5593.485  7364.584
#>   [20,] 4104.771  7798.971
#>   [21,] 4751.687  6173.084
#>   [22,] 6880.196  7446.442
#>   [23,] 4193.665  6759.968
#>   [24,] 7894.584  5499.272
#>   [25,] 6790.802  6755.698
#>   [26,] 5773.908  6958.576
#>   [27,] 5025.982  7949.259
#>   [28,] 5423.312  8156.845
#>   [29,] 4477.059  7119.983
#>   [30,] 5248.368  5984.532
#>   [31,] 4118.866  5167.377
#>   [32,] 4865.489  6002.129
#>   [33,] 5901.482  7333.660
#>   [34,] 2870.309  5249.494
#>   [35,] 4296.687  7034.716
#>   [36,] 5667.415  5629.202
#>   [37,] 6088.682  6649.954
#>   [38,] 5331.251  5467.776
#>   [39,] 4508.196  7688.174
#>   [40,] 6196.258  5456.295
#>   [41,] 4004.092  7024.475
#>   [42,] 5997.418  6972.999
#>   [43,] 5681.549  6219.739
#>   [44,] 4567.018  7466.110
#>   [45,] 6485.303  7941.202
#>   [46,] 4810.985  8222.171
#>   [47,] 4505.531  5358.426
#>   [48,] 4571.070  6419.743
#>   [49,] 7197.847  7771.455
#>   [50,] 4853.474  6545.940
#>   [51,] 4859.328  7163.862
#>   [52,] 4850.774  7486.495
#>   [53,] 3824.862  6249.897
#>   [54,] 6154.534  4227.242
#>   [55,] 4339.029  6951.002
#>   [56,] 4874.001  6910.913
#>   [57,] 6181.824  5724.119
#>   [58,] 4890.276  9126.803
#>   [59,] 5946.322  5641.182
#>   [60,] 4546.802  7082.214
#>   [61,] 2888.008  7532.775
#>   [62,] 6047.198  5928.406
#>   [63,] 3710.438  7562.438
#>   [64,] 4109.935  6294.740
#>   [65,] 3073.127  7445.990
#>   [66,] 5817.619  8485.683
#>   [67,] 3704.948  7078.660
#>   [68,] 4483.458  7092.427
#>   [69,] 5465.723  7951.119
#>   [70,] 4708.959  7792.106
#>   [71,] 6860.363  5125.030
#>   [72,] 6222.054  6620.798
#>   [73,] 2743.096  6047.325
#>   [74,] 4868.001  8952.385
#>   [75,] 4638.767  5578.560
#>   [76,] 4462.378  8298.910
#>   [77,] 5526.927  6377.898
#>   [78,] 4829.884  7330.658
#>   [79,] 3913.710  6464.886
#>   [80,] 4549.977  6517.369
#>   [81,] 5764.026  6792.898
#>   [82,] 4196.409  6571.181
#>   [83,] 7219.244  7184.631
#>   [84,] 4402.528  6662.141
#>   [85,] 5791.254  9798.311
#>   [86,] 6512.304  7838.256
#>   [87,] 5383.137  7607.833
#>   [88,] 3479.228  5266.471
#>   [89,] 4806.621  7687.010
#>   [90,] 5536.833  9931.489
#>   [91,] 4529.517  8074.569
#>   [92,] 5496.123  8233.682
#>   [93,] 5326.235  8401.309
#>   [94,] 4684.898  6319.692
#>   [95,] 4089.152  8318.027
#>   [96,] 5380.264  6719.110
#>   [97,] 4237.119  4823.627
#>   [98,] 5745.763  6840.280
#>   [99,] 5569.928  6751.252
#>  [100,] 5479.220  6083.096
#>  [101,] 4540.762  5683.469
#>  [102,] 4819.641  6230.629
#>  [103,] 4253.110  7319.877
#>  [104,] 4774.110  6087.789
#>  [105,] 4359.650  8300.852
#>  [106,] 5753.817  5363.774
#>  [107,] 5923.174  6270.048
#>  [108,] 5147.923  7515.306
#>  [109,] 4421.472  6186.464
#>  [110,] 6654.006  5948.871
#>  [111,] 5539.179  7209.449
#>  [112,] 6149.176  6306.154
#>  [113,] 4753.325  8765.393
#>  [114,] 4999.901  7032.875
#>  [115,] 6252.385  8328.264
#>  [116,] 6034.705  5224.194
#>  [117,] 5281.030  7453.252
#>  [118,] 4593.627  5719.098
#>  [119,] 3700.205  6569.601
#>  [120,] 5981.044  6030.773
#>  [121,] 4712.294  5698.198
#>  [122,] 5028.167  8419.706
#>  [123,] 3972.976  4960.548
#>  [124,] 5836.381  4556.311
#>  [125,] 3366.761  6990.169
#>  [126,] 4372.133  5363.149
#>  [127,] 4931.896  7084.041
#>  [128,] 3419.483  5382.083
#>  [129,] 3724.859  6519.627
#>  [130,] 6049.545  7676.144
#>  [131,] 5766.301  6486.135
#>  [132,] 5146.857  6965.385
#>  [133,] 3886.262  7641.638
#>  [134,] 6509.783  7320.719
#>  [135,] 4650.430  9116.937
#>  [136,] 3584.470  8026.941
#>  [137,] 5513.665  6558.771
#>  [138,] 5284.960  8593.298
#>  [139,] 5922.250  7000.727
#>  [140,] 5375.979  5874.566
#>  [141,] 3295.660  6229.739
#>  [142,] 4199.805  5340.496
#>  [143,] 4178.662  6943.381
#>  [144,] 4862.087  5624.589
#>  [145,] 5578.652  6637.472
#>  [146,] 4798.041  8794.992
#>  [147,] 4598.147  5451.016
#>  [148,] 5872.558  5566.967
#>  [149,] 6210.700  5718.127
#>  [150,] 3758.558  6267.847
#>  [151,] 5737.963  6765.737
#>  [152,] 3292.035  5697.534
#>  [153,] 4592.052  9486.056
#>  [154,] 4181.267  6834.661
#>  [155,] 5458.441  7488.704
#>  [156,] 4088.115  7458.458
#>  [157,] 7209.003  5437.035
#>  [158,] 4244.064  4653.845
#>  [159,] 4765.500  7664.829
#>  [160,] 5748.063  6870.556
#>  [161,] 4775.744  6504.365
#>  [162,] 3027.804  5443.462
#>  [163,] 4897.787  7140.380
#>  [164,] 4527.583  5484.684
#>  [165,] 5194.358  7037.609
#>  [166,] 5625.542  9441.576
#>  [167,] 5799.473  5852.138
#>  [168,] 4565.032  8550.785
#>  [169,] 4204.356  6765.780
#>  [170,] 5183.331  5695.332
#>  [171,] 7806.275  7405.615
#>  [172,] 4625.100  7161.118
#>  [173,] 4588.586  6986.705
#>  [174,] 2925.049  7101.557
#>  [175,] 5326.961  7387.459
#>  [176,] 4146.804  7020.944
#>  [177,] 4270.597  6325.790
#>  [178,] 5843.599  6840.949
#>  [179,] 4947.082  6361.894
#>  [180,] 5461.752  7332.159
#>  [181,] 6887.878  7989.378
#>  [182,] 3350.770  6516.694
#>  [183,] 4901.015  8067.435
#>  [184,] 4739.859  7745.663
#>  [185,] 5588.445  6831.972
#>  [186,] 4587.833  8559.114
#>  [187,] 5338.759  5515.927
#>  [188,] 5143.068  7132.702
#>  [189,] 4272.088  5965.390
#>  [190,] 3039.558  6346.339
#>  [191,] 5380.333  6822.393
#>  [192,] 5718.584  7226.525
#>  [193,] 5837.820  6395.248
#>  [194,] 4114.424  9027.259
#>  [195,] 5356.493  6242.138
#>  [196,] 3174.006  9036.378
#>  [197,] 6222.604  5517.819
#>  [198,] 6765.514  8500.395
#>  [199,] 5328.736  6734.005
#>  [200,] 6148.602  4808.223
#>  [201,] 6467.094  6525.636
#>  [202,] 7271.251  5987.755
#>  [203,] 5672.226  7082.377
#>  [204,] 4587.545  7219.762
#>  [205,] 5341.423  8413.402
#>  [206,] 3738.046  7529.313
#>  [207,] 5354.675  6244.271
#>  [208,] 6079.065 10073.910
#>  [209,] 3267.374  5716.259
#>  [210,] 6102.888  6619.542
#>  [211,] 7472.955  5442.807
#>  [212,] 5469.325  9105.947
#>  [213,] 6294.438  6253.524
#>  [214,] 5571.531  8173.788
#>  [215,] 5428.494  8296.548
#>  [216,] 5303.874  5774.271
#>  [217,] 4043.944  9754.518
#>  [218,] 4737.978  6629.495
#>  [219,] 4738.951  8018.219
#>  [220,] 5193.288  5982.292
#>  [221,] 5374.517  7407.635
#>  [222,] 4139.739  5876.436
#>  [223,] 6423.733  6588.779
#>  [224,] 5599.179  5844.191
#>  [225,] 6388.091  6878.330
#>  [226,] 4500.137  7077.421
#>  [227,] 4397.929  6720.973
#>  [228,] 4152.265  7973.242
#>  [229,] 3645.614  6336.914
#>  [230,] 5167.528  6643.790
#>  [231,] 4487.177  6805.549
#>  [232,] 4139.563  5602.049
#>  [233,] 6759.578  7860.936
#>  [234,] 4156.601  8929.465
#>  [235,] 4610.915  6473.340
#>  [236,] 3547.304  7617.901
#>  [237,] 4874.972  5943.236
#>  [238,] 4318.732  6610.623
#>  [239,] 4049.616  7489.533
#>  [240,] 4846.061  9342.601
#>  [241,] 5302.093  7730.159
#>  [242,] 4735.766  6623.806
#>  [243,] 5170.683  5948.673
#>  [244,] 4530.013  7713.664
#>  [245,] 5338.428  8029.578
#>  [246,] 3635.621 10279.058
#>  [247,] 6019.484  5564.355
#>  [248,] 6187.934  5703.801
#>  [249,] 5886.401  6984.397
#>  [250,] 5578.873  5959.188
#>  [251,] 5350.228  7084.823
#>  [252,] 3821.051  6739.825
#>  [253,] 5218.062  6784.457
#>  [254,] 3621.708  6125.744
#>  [255,] 5094.548  6364.944
#>  [256,] 5391.656  5551.171
#>  [257,] 7092.025  7276.001
#>  [258,] 4098.141  6153.487
#>  [259,] 3548.500  6465.829
#>  [260,] 6893.320  6654.054
#>  [261,] 6296.068  8654.136
#>  [262,] 5380.902  6425.963
#>  [263,] 4299.317 11395.643
#>  [264,] 4563.655  7501.372
#>  [265,] 6179.188  6272.991
#>  [266,] 5904.668  8177.355
#>  [267,] 2990.018  9547.293
#>  [268,] 4177.868  6178.459
#>  [269,] 3507.798  5271.042
#>  [270,] 7137.089  9458.619
#>  [271,] 4408.303  5399.935
#>  [272,] 4298.816  6138.675
#>  [273,] 5653.141  7596.384
#>  [274,] 4748.853  5834.563
#>  [275,] 6486.633  7219.549
#>  [276,] 4597.355  6837.977
#>  [277,] 3569.383  6638.205
#>  [278,] 3542.726  5615.692
#>  [279,] 2958.642  7294.211
#>  [280,] 4898.019  6718.248
#>  [281,] 6284.097  6059.799
#>  [282,] 4596.594  7806.713
#>  [283,] 4531.774  7014.111
#>  [284,] 4383.017  7599.817
#>  [285,] 3801.688  8521.861
#>  [286,] 7100.422  6472.806
#>  [287,] 5400.682  4696.277
#>  [288,] 4484.011  5404.355
#>  [289,] 5571.654  8004.154
#>  [290,] 5314.027  6853.119
#>  [291,] 7161.231  5489.053
#>  [292,] 4502.787  5140.552
#>  [293,] 4410.923  6869.224
#>  [294,] 5881.371  4588.006
#>  [295,] 5319.721  5565.961
#>  [296,] 5127.616  8613.681
#>  [297,] 5443.237  8035.442
#>  [298,] 4583.404  4928.697
#>  [299,] 4739.743  6945.141
#>  [300,] 3794.536  4661.775
#>  [301,] 5224.993  6924.280
#>  [302,] 4656.190  8140.862
#>  [303,] 4429.490  4926.965
#>  [304,] 4116.804  7890.956
#>  [305,] 4246.440  7404.236
#>  [306,] 5474.046  5900.131
#>  [307,] 4991.837  6484.581
#>  [308,] 4617.468  8784.330
#>  [309,] 5390.439  4676.296
#>  [310,] 6471.935  8997.444
#>  [311,] 4723.436  7549.632
#>  [312,] 4409.631  8784.430
#>  [313,] 4713.871  6514.419
#>  [314,] 3935.375  6516.209
#>  [315,] 4848.417  7425.818
#>  [316,] 7133.660  5764.946
#>  [317,] 5158.203  7139.029
#>  [318,] 5388.852  7513.824
#>  [319,] 4392.472  9442.285
#>  [320,] 5409.075  6825.239
#>  [321,] 5066.369  6286.777
#>  [322,] 3438.624  7222.827
#>  [323,] 4741.971  9230.137
#>  [324,] 5328.750  7823.918
#>  [325,] 6259.260  8167.590
#>  [326,] 5242.696  7019.771
#>  [327,] 4028.704  6778.851
#>  [328,] 5730.668  9746.002
#>  [329,] 3314.874  7619.273
#>  [330,] 5715.952  6989.013
#>  [331,] 4833.075  7703.472
#>  [332,] 6124.622  5716.904
#>  [333,] 5259.200  7403.963
#>  [334,] 5490.158  5251.919
#>  [335,] 6078.713  6470.282
#>  [336,] 3861.088  5687.205
#>  [337,] 6064.838  7462.583
#>  [338,] 3899.209  6351.351
#>  [339,] 4004.388  5912.957
#>  [340,] 3744.798  6602.771
#>  [341,] 4858.874  8126.005
#>  [342,] 6019.752  5468.542
#>  [343,] 5304.111  7920.392
#>  [344,] 4602.586  7679.018
#>  [345,] 5041.099  6812.481
#>  [346,] 4237.779  8755.906
#>  [347,] 5836.355  6319.537
#>  [348,] 4255.775  6844.869
#>  [349,] 3827.005  6475.041
#>  [350,] 6682.657  8358.500
#>  [351,] 5691.784  6704.216
#>  [352,] 4408.973  8765.121
#>  [353,] 3512.947  7481.046
#>  [354,] 5082.097  8749.950
#>  [355,] 3526.651  7320.607
#>  [356,] 5679.338  7248.036
#>  [357,] 4922.656  6651.441
#>  [358,] 4420.798  6477.740
#>  [359,] 5483.573  5145.096
#>  [360,] 5419.862  5644.434
#>  [361,] 5589.642  8488.436
#>  [362,] 4886.178  6334.795
#>  [363,] 4365.057  8962.848
#>  [364,] 4794.209  5414.390
#>  [365,] 5740.777  9770.894
#>  [366,] 4258.213  7545.006
#>  [367,] 5661.947  6707.575
#>  [368,] 4235.756  5469.978
#>  [369,] 3864.529  6590.262
#>  [370,] 5688.447  7739.683
#>  [371,] 6752.291  7948.452
#>  [372,] 5334.829  5697.419
#>  [373,] 4280.993  6521.862
#>  [374,] 3861.259  7206.193
#>  [375,] 3003.690  8385.206
#>  [376,] 5321.613  7131.548
#>  [377,] 5316.474  7624.974
#>  [378,] 4819.889  6962.412
#>  [379,] 5267.202  5579.052
#>  [380,] 5364.321  6391.514
#>  [381,] 6167.244  6392.184
#>  [382,] 5046.372  7716.955
#>  [383,] 4356.375  6296.188
#>  [384,] 6920.794  6306.251
#>  [385,] 3967.892  7959.148
#>  [386,] 4096.795  5739.375
#>  [387,] 4942.394  7584.422
#>  [388,] 4285.377  6062.710
#>  [389,] 5084.679  7952.637
#>  [390,] 3003.184  7498.883
#>  [391,] 3505.232  7036.097
#>  [392,] 4166.140  5282.424
#>  [393,] 7721.057  7559.312
#>  [394,] 6440.087  5762.113
#>  [395,] 4917.848  5868.496
#>  [396,] 6142.202  4839.029
#>  [397,] 4198.831  8179.579
#>  [398,] 4538.166  4808.146
#>  [399,] 3345.050  6956.882
#>  [400,] 3919.861  7412.088
#>  [401,] 5344.516  9005.052
#>  [402,] 5381.194  7010.934
#>  [403,] 6226.396  7378.516
#>  [404,] 4598.210  7163.059
#>  [405,] 5319.064  8144.278
#>  [406,] 8221.917  8114.185
#>  [407,] 3954.550  6126.552
#>  [408,] 4552.725  7507.572
#>  [409,] 4941.024  6834.059
#>  [410,] 5882.206  8373.761
#>  [411,] 6344.818  7560.286
#>  [412,] 5643.597  7766.169
#>  [413,] 5921.768  7638.682
#>  [414,] 5456.196  7718.106
#>  [415,] 4470.763  7766.456
#>  [416,] 4436.307  6486.253
#>  [417,] 4639.957  5197.727
#>  [418,] 5866.314  6143.737
#>  [419,] 4465.344  8491.078
#>  [420,] 5319.886  6839.962
#>  [421,] 5245.219  6051.157
#>  [422,] 4762.952  7448.211
#>  [423,] 4000.123  5519.648
#>  [424,] 5278.765  7330.723
#>  [425,] 5591.624  8029.797
#>  [426,] 4496.118  8515.986
#>  [427,] 4690.921  6256.161
#>  [428,] 4454.708  5839.614
#>  [429,] 3822.905  7335.175
#>  [430,] 5618.483  8722.608
#>  [431,] 6092.593  7102.072
#>  [432,] 4892.915  4416.428
#>  [433,] 4292.630 10232.974
#>  [434,] 5120.640  7679.164
#>  [435,] 5491.404  5493.712
#>  [436,] 5165.386  7695.063
#>  [437,] 4494.634  6959.822
#>  [438,] 6160.868  6543.484
#>  [439,] 4932.086  6318.935
#>  [440,] 7494.062  6687.832
#>  [441,] 5435.475  8493.894
#>  [442,] 4335.576  6812.295
#>  [443,] 4974.402  4632.287
#>  [444,] 5166.468  7092.928
#>  [445,] 4018.399  7340.330
#>  [446,] 4724.917  5450.274
#>  [447,] 4911.606  3984.621
#>  [448,] 4828.705  5727.171
#>  [449,] 5716.818  7837.987
#>  [450,] 6199.924  6365.652
#>  [451,] 5158.081  6789.218
#>  [452,] 6288.596  8505.100
#>  [453,] 6181.683  4342.138
#>  [454,] 5355.799  7366.246
#>  [455,] 4647.637  6488.567
#>  [456,] 5232.028  7104.658
#>  [457,] 5331.455  5961.955
#>  [458,] 4668.517  7369.639
#>  [459,] 5892.875  6696.164
#>  [460,] 4241.368  6068.092
#>  [461,] 6297.216  6072.485
#>  [462,] 4597.007  7879.357
#>  [463,] 4942.838  6428.780
#>  [464,] 3909.759  6763.849
#>  [465,] 5580.711  9817.794
#>  [466,] 4322.358  5844.526
#>  [467,] 7117.670  7418.071
#>  [468,] 5735.754  6691.088
#>  [469,] 5693.128  6216.618
#>  [470,] 3868.261  5369.317
#>  [471,] 4600.343  9634.189
#>  [472,] 3066.642  7299.487
#>  [473,] 3648.364  7894.364
#>  [474,] 5228.306  5016.895
#>  [475,] 3611.539  6616.964
#>  [476,] 5244.307  5732.977
#>  [477,] 6212.710  6101.529
#>  [478,] 4146.984  6650.165
#>  [479,] 5894.857  9358.543
#>  [480,] 5625.327  5325.368
#>  [481,] 5789.757  5707.247
#>  [482,] 6216.412  7527.264
#>  [483,] 4070.054  8994.160
#>  [484,] 5739.548  7914.990
#>  [485,] 4563.679  7869.068
#>  [486,] 5295.905  7434.767
#>  [487,] 5564.901  7196.171
#>  [488,] 5944.666  5843.811
#>  [489,] 4449.994  6749.791
#>  [490,] 5166.208  7654.255
#>  [491,] 5377.442  7459.849
#>  [492,] 4742.731  6488.526
#>  [493,] 5772.585  6221.578
#>  [494,] 6164.413  8282.827
#>  [495,] 4396.780  6720.774
#>  [496,] 5307.828  8629.578
#>  [497,] 4857.938  6801.057
#>  [498,] 5703.106  7192.238
#>  [499,] 5929.955  7721.462
#>  [500,] 4676.963  6802.201
#>  [ reached getOption("max.print") -- omitted 500 rows ]
#> 
#> $timefun
#> NULL
#> 
#> attr(,"class")
#> [1] "params_statevals"