Parameters of a 2D nested list of survival models

Create a two-dimensional nested list containing the parameters of multiple fitted parametric survival models. This would typically be used for joining multiple survival models at different time points with join.

params_surv_list_2d(coefs, dists)

Arguments

coefs
A list of lists of lists. The number of lists is equal to the number of fitted models. The outer most list, or 1st list, contains survival models used to predict different outcomes (e.g., curves in a partitioned survival analysis or states in a multi-state model). The 2nd list contains multiple specifications of that model (e.g., for joining multiple survival models at different points in time). Finally, the 3rd, or inner most list, is a list of the parameters in a given survival model as specified in params_surv.

dists
A list of character vectors, where each element of the list contains a vector of the parametric distributions used to predict a given survival outcome. The allowed distributions are detailed in params_surv.

Value

An object of class "params_surv_list_2d", which is a list containing coefs, dists, and n_sims. n_sims is equal to the number of rows in each element of coefs, which must be the same across parameters and fitted survival models.