Personalized cost-effectiveness analysis

Conduct Bayesian cost-effectiveness analysis (e.g. summarize a probabilistic sensitivity analysis) by subgroup.

pcea(x, k, sim = "sim", arm = "arm", grp = "grp", e = "e", c = "c",
  custom_vars = NULL, custom_fun = NULL)

pcea_pw(x, k, control, sim = "sim", arm = "arm", grp = "grp", e = "e",
  c = "c", custom_vars = NULL, custom_fun = NULL)

Arguments

x

Matrix containing information on mean costs and effectiveness for each simulation. Should be in long form with unit of observation as the simulation and treatment arm. Should have the following columns ( sim = simulation number, arm = treatment arm, c = summary of cost for each simulation and treatment arm e = summary of clinicial effectiveness for each simulation and treatment arm)

k

Vector of willingess to pay values

sim

Name of column denoting simulation number. Default is "sim".

arm

Name of column denoting treatment arm. Default is "arm".

grp

Name of column denoting subgroup. Default is "grp".

e

Name of column denoting clinical effectiveness. Default is "e".

c

Name of column denoting costs. Default is "c".

custom_vars

Character vector of variable names to use in creating a custom summary table. Table will contain means and 95% credible intervals for each variable. Can contain e and c.

custom_fun

Function to apply to custom_vars to make custom table. If custom_vars is not NULL and custom_fun is NULL, then returns the mean, 2.5% quantile, and 97.5% quantile for each variable in custom_vars.

control

Name of the control arm in x.

Value

pcea returns a list containing four data.tables:

summary

A data.table of the mean, 2.5% quantile, and 97.5% quantile by arm and group for clinical effectiveness and costs.

mce

The probability that each arm is the most effective treatment for each group for the range of specified willingess to pay values.

evpi

The expected value of perfect information by group for the range of specified willingess to pay values.

nb

The mean, 2.5% quantile, and 97.5% quantile of (monetary) net benefits for the range of specified willingess to pay values.

In addition, if custom_vars is not NULL, pcea returns custom.table, which is a data.table containing summary statistics for each variable in custom_vars by arm and group.

pcea_pw also returns a list containing four data.tables:

summary

A data.table of the mean, 2.5% quantile, and 97.5% quantile by arm and group for clinical effectiveness and costs.

delta

Incremental effectiveness and incremental cost for each simulated parameter set by arm and group. Can be used to plot a cost-effectiveness plane. Also returns the difference between each treatment arm and the comparator for each variable in custom_vars if custom_vars is not NULL.

ceac

Values needed to plot a cost-effectiveness acceptability curve by group. In other words, the probability that each arm is more cost-effective than the comparator for the specified willingess to pay values.

inb

The mean, 2.5% quantile, and 97.5% quantile of (monetary) incremental net benefits for the range of specified willingess to pay values.

If custom_vars is not NULL, also returns custom.table, which is a data.table containing summary statistics for the values of each variable returned in delta by arm and group.