Plot cost-effectiveness plane

Plot a cost-effectiveness plane from the output of cea_pw() using ggplot2. Each point is a random draw of incremental costs (y-axis) and incremental QALYs (x-axis) from a probabilistic sensitivity analysis.

plot_ceplane(x, k = 50000, labels = NULL)

Arguments

x
A cea_pw object produced by cea_pw().

k
Willingness to pay per QALY.

labels
A list of named vectors containing the values and labels of variables. The elements of each vector are the values of a variable and the names are the labels. The names of the list are the names of the variables. See the output returned by get_labels() for an example.

Value

A ggplot object.

Details

See the cea_pw() documentation for an example. If there are multiple subgroups, then a faceted plot is produced with one plot for each subgroup.