Quality-adjusted life-years object

An object of class qalys returned from methods $sim_qalys() in model classes that store simulated quality-adjusted life-years (QALYs).

Components

A qalys object inherits from data.table and contains the following columns:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

The patient ID.

state_id

The health state ID.

dr

The rate used to discount QALYs.

category

A single category always equal to "qalys".

qalys

The simulated values of QALYs.

If the argument lys = TRUE, then the data.table also contains a column lys containing simulated life-years.