Randomly sample parameters from a flexsurvreg model

Draw random samples from a survival model fit using flexsurv::flexsurvreg or flexsurv::flexsurvspline.

# S3 method for flexsurvreg
rsample(x, n = NULL, point_estimate = FALSE, ...)

Arguments

x
Output from flexsurvreg from the package flexsurv.

n
Number of random observations to draw.

point_estimate
If TRUE, then the point estimate is used for prediction and no samples are drawn.

...
Further arguments passed to or from other methods. Currently unused.

Value

Returns an object of class "survlist_pars". A list with the following components is returned:

coefs

List where each element is a matrix of sampled values of regression coefficients used to predict a parameter of the probability distribution.

dist

Name of the probability distribution.

Examples

library("flexsurv")
fit <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1,
                   data = ovarian, dist = "weibull")
n <- 5
fit.sample <- rsample(fit, n = n)
head(fit.sample$coefs)

#> NULL

fit.sample$dist

#> NULL