Randomly sample parameters from a spline-based survival model

Draw random samples from a survival model fit using flexsurv::flexsurvspline.

# S3 method for flexsurvspline
rsample(x, n, ...)

Arguments

x
Output from flexsurvspline from the package flexsurv.

n
Number of random observations to draw.

...
Further arguments passed to or from other methods. Currently unused.

Value

List where each element contains random samples of the coefficients used to predict a given parameter of the survival distribution.

Examples

library("flexsurv")
fit <- flexsurv::flexsurvspline(Surv(recyrs, censrec) ~ group, data = bc, k=1,
                     scale = "hazard")
fit.sample <- rsample(fit, n = 2)
head(fit.sample)

#> $gamma0
#>         gamma0 groupMedium groupPoor
#> [1,] -2.999321   0.5893955  1.511952
#> [2,] -3.261121   0.8858128  1.660297
#> 
#> $gamma1
#>        gamma1
#> [1,] 2.970830
#> [2,] 2.599603
#> 
#> $gamma2
#>         gamma2
#> [1,] 0.2524031
#> [2,] 0.1769695
#>