Analyze simulated multi-state model

Analyze a simulated multi-state model

sim_pv(sim, tmat, r = 0.03, x = NULL, beta, poly.beta = NULL,
  poly.deg = NULL, knots = NULL, name = NULL, agevar = NULL)

sim_los(sim, timevar)

sim_transprob(sim, t)

Arguments

sim

A simulated multi-state model from sim_msm.

r

Discount rate for the model. Default is 0.03. Implies a discount factor of \(1/(1 + r)^t\) where t is the cycle number.

x

Data matrix

beta

Matrix of regression coefficients corresponding to data matrix x. One row for each state.

poly.beta

Matrix of coefficeints for piecewise time polynomials. One row for each state.

poly.deg

Matrix indicating degree of polynomial on each interval. One row for each state.

knots

Matrix indicating location of knots for piecewise polynomials. One row for each state.

Value

Vector of present value of quantity for each simulated individual during time spent in state.