Simulate design matrix

Simulate a design matrix for regression analysis consiting of a binary variable for female gender and a continuous variable for age. Gender is drawn from a binomial distribution and age from a lognormal distribution.

sim_design_matrix(n_obs = 10, female_prop = 0.51, age_meanlog = 4,
  age_sdlog = 0.25)

Arguments

n_obs
Number of observations to sample.

female_prop
Proportion of the population that is female.

age_meanlog
Mean of age on the log scale.

age_sdlog
Standard deviation of age on the log scale.