Mean values for health states

Estimated means that are used to simulate values assigned to health states with StateVals.

stateval_means(values, strategy_id, patient_id)

Arguments

values
Typically a matrix where each column denotes a health state and each row denotes a random sample of the value assigned to that health state for the probabilistic sensitivitiy analysis. If health state values vary by strategy, then it can be an array of matrices where each matrix denotes the values associated with a different treatment strategy.

strategy_id
If a matrix, a vector denoting the strategy ID's to be modeled; if an array, a vector denoting the strategy ID associated with each matrix in the array.

patient_id
The patient ID's that will be modeled in the simulation. Note that state values are assumed to be constant across patients when using objects of class "stateval_means".

Value

An object of class "stateval_means", which is a list containing values, strategy_id, and patient_id.

Examples

# Cost estimates in 2 health states for a model with 2 treatment strategies and 3 patients
gamma_params <- mom_gamma(c(5000, 7000), c(1000, 1200))
n <- 3
vals <- matrix(rgamma(2 * n,
                      shape = gamma_params$shape,
                      scale = gamma_params$scale),
               nrow = n, ncol = 2, byrow = TRUE)
stval_ests <- stateval_means(values = vals,
                           strategy_id = c(1, 2),
                           patient_id = c(1, 2, 3))
print(stval_ests)

#> $values
#>          [,1]     [,2]
#> [1,] 4018.652 7062.365
#> [2,] 7267.440 7270.911
#> [3,] 4348.291 6618.499
#> 
#> $strategy_id
#> [1] 1 2
#> 
#> $patient_id
#> [1] 1 2 3
#> 
#> attr(,"class")
#> [1] "stateval_means"

stateval_mod <- create_StateVals(stval_ests)
head(stateval_mod$sim(t = c(1, 2, 3), type = "predict"))

#>    sample strategy_id patient_id state_id time    value
#> 1:      1           1          1        1    1 4018.652
#> 2:      1           1          1        2    1 7062.365
#> 3:      1           1          2        1    1 4018.652
#> 4:      1           1          2        2    1 7062.365
#> 5:      1           1          3        1    1 4018.652
#> 6:      1           1          3        2    1 7062.365