Summarize eval_rng object

Summarize the model parameters randomly sampled for probabilistic sensitivity analysis with eval_rng().

# S3 method for eval_rng
summary(object, probs = c(0.025, 0.975), sep = "_", ...)

# S3 method for eval_rng
print(x, ...)

Arguments

object, x
An eval_rng object.

probs
A numeric vector of probabilities with values in [0,1] used to compute quantiles with stats::quantile().

sep
When a list element returned by eval_rng is a tabular object, the parameter name is created by concatenating the name of the list element with the columns of the tabular object. The sep argument determines the character string used to separate the terms.

...
For the print method, arguments to pass to summary.eval_rng().

Value

summary.eval_rng() returns a data.table with columns for (i) the name of the parameter (param), (ii) the mean of the parameter samples (mean), (iii) the standard deviation of the parameter samples (sd), and (iv) quantiles of the parameter samples corresponding to the probs argument. print.eval_rng() prints the output of summary.eval_rng() to the console.

See also