Summarize tparams_transprobs object

The summary() method summarizes a tparams_transprobs object containing predicted transition probabilities; summary statistics are computed for each possible transition by the relevant ID variables.

# S3 method for class 'tparams_transprobs'
summary(object, probs = NULL, unflatten = FALSE, ...)

Arguments

object

A tparams_transprobs object.

probs

A numeric vector of probabilities with values in [0,1] used to compute quantiles. Computing quantiles can be slow when object is large, so the default is NULL, meaning that no quantiles are computed.

unflatten

If FALSE, then each column containing a summary statistic is a vector and the generated table contains one row (for each set of ID variables) for each possible transition; if TRUE, then each column stores a list of matrix objects containing transition probability matrices formed by "unflattening" the one-dimensional vectors. See "Value" below for additional details.

...

Additional arguments affecting the summary. Currently unused.

Value

If unflatten = "FALSE" (the default), then a data.table::data.table is returned with columns for (i) the health state that is being transitioned from (from), (ii) the health state that is being transitioned to (to) (iii) the mean of each parameter across parameter samples (mean), (iv) the standard deviation of the parameter samples (sd), and (v) quantiles of the parameter samples corresponding to the probs argument.

If, on the other hand, unflatten = "TRUE", then the parameters are unflattened to form transition probability matrices; that is, the mean, sd, and quantile columns are (lists of) matrices.

In both cases, the ID variables are also returned as columns.

See also