Table to store transition probabilities

Create a table for storing transition probabilities for a discrete time state transition model stratified by treatment strategy, patient, transition, and (optionally) time interval.

transprob_tbl(tbl, dist = c("custom", "dirichlet", "fixed"),
  hesim_data = NULL)

Arguments

tbl
A data.frame or data.table for storing parameter values. See "Details" for specifics.

dist
Probability distribution used to sample parameters for a probabilistic sensitivity analysis (PSA).

hesim_data
A hesim_data object. Required to specify patients if not included as columns in tbl. Not required if tbl includes one row for each treatment strategy, patient, and transition combination.

Value

An object of class "transprob_tbl", which is a data.table of parameter values with attributes for dist and optionally patients. tbl is in the same format as described in "Details". patients is a data.table with at leat one column containing patient_id.

Details

tbl is a data.table that must contain columns for treatment strategies (strategy_id) and health state transitions (transition_id). The number of total transitions must be equal to the square of number of health states. The transition IDs are, by default, integers starting at 1 and ordered rowwise in transition matrices where each (i,j) pair denotes a transition from state i to state j. For example, in a model with two health states transition 1 would be the transition from state 1 to itself, transition 2 would be the transition from state 1 to state 2, transition 3 would be the transition from state 2 to state 1, and transition 4 would be the transition from state 2 to itself. Alternatively, transition IDs can be specified in hesim_data by including a transitions table. Each row is a treatment strategy and transition pair.

Transition probabilities may also vary by patient (patient_id) and time intervals (time_start). If both of these variables are included, then each row is a treatment strategy, transition, patient, and time interval pair. If no patient ID is specified in the table, then the patients in the simulation can be specified using the hesim_data argument and it will be assumed that transition probabilities are the same across patients.

tbl must also contain columns summarizing the state values for each row. Parameter values for a PSA can be drawn from Dirichlet distributions (separate for each treatment strategy and optionally patient subgroup and time interval) with dist = 'dirichlet'. Alternatively, fixed can be used if estimates are known with certainty and custom can be used if transition probabilities for a PSA have been previously sampled from an arbitrary probability distribution. If dist = "custom", then tbl must include a column named sample (an integer vector denoting a unique random draw). The columns in tbl related to the values of the transition probabilities depend on the value of dist. They are as follows:

fixed

est

dirichlet

alpha

custom

value

See also