Cohort discrete time state transition model

Simulate outcomes from a cohort discrete time state transition model.

Format

An R6::R6Class object.

See also

Public fields

trans_model

The model for health state transitions. Must be an object of class CohortDtstmTrans.

utility_model

The model for health state utility. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

stateprobs_

An object of class stateprobs simulated using $sim_stateprobs().

qalys_

An object of class qalys simulated using $sim_qalys().

costs_

An object of class costs simulated using $sim_costs().

Methods

Public methods

CohortDtstm$new()

CohortDtstm$sim_stateprobs()

CohortDtstm$sim_qalys()

CohortDtstm$sim_costs()

CohortDtstm$summarize()

CohortDtstm$clone()

Method new()
Create a new CohortDtstm object.
Usage

CohortDtstm$new(trans_model = NULL, utility_model = NULL, cost_models = NULL)

Arguments

trans_model

The trans_model field.

utility_model

The utility_model field.

cost_models

The cost_models field.

Returns
A new CohortDtstm object.

Method sim_stateprobs()
Simulate health state probabilities using CohortCtstmTrans$sim_stateprobs().
Usage

CohortDtstm$sim_stateprobs(n_cycles)

Arguments

n_cycles

The number of model cycles to simulate the model for.

Returns
An instance of self with simulated output of class stateprobs stored in stateprobs_.

Method sim_qalys()
Simulate quality-adjusted life-years (QALYs) using sim_qalys().
Usage

CohortDtstm$sim_qalys(
  dr = 0.03,
  method = c("trapz", "riemann_left", "riemann_right"),
  lys = FALSE
)

Arguments

dr

Discount rate.

method

Method used to integrate state values when computing (QALYs).

lys

If TRUE, then life-years are simulated in addition to QALYs.

Returns
An instance of self with simulated output of class qalys stored in qalys_.

Method sim_costs()
Simulate costs using sim_costs().
Usage

CohortDtstm$sim_costs(
  dr = 0.03,
  method = c("trapz", "riemann_left", "riemann_right")
)

Arguments

dr

Discount rate.

method

Method used to integrate state values when computing costs.

Returns
An instance of self with simulated output of class costs stored in costs_.

Method summarize()
Summarize costs and QALYs so that cost-effectiveness analysis can be performed. See summarize_ce().
Usage

CohortDtstm$summarize()

Method clone()
The objects of this class are cloneable with this method.
Usage

CohortDtstm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.