Cohort discrete time state transition model

Simulate outcomes from a cohort discrete time state transition model.

Format

An R6::R6Class object.

See also

Public fields

trans_model

The model for health state transitions. Must be an object of class CohortDtstmTrans.

utility_model

The model for health state utility. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

stateprobs_

An object of class stateprobs simulated using $sim_stateprobs().

qalys_

An object of class qalys simulated using $sim_qalys().

costs_

An object of class costs simulated using $sim_costs().

Methods

Public methods

CohortDtstm$new()

CohortDtstm$sim_stateprobs()

CohortDtstm$sim_qalys()

CohortDtstm$sim_costs()

CohortDtstm$summarize()

CohortDtstm$clone()

Method new()

Create a new CohortDtstm object.

Usage

CohortDtstm$new(trans_model = NULL, utility_model = NULL, cost_models = NULL)

Arguments

trans_model

The trans_model field.

utility_model

The utility_model field.

cost_models

The cost_models field.

Returns

A new CohortDtstm object.

Method sim_stateprobs()

Simulate health state probabilities using CohortDtstmTrans$sim_stateprobs().

Usage

CohortDtstm$sim_stateprobs(n_cycles)

Arguments

n_cycles

The number of model cycles to simulate the model for.

Returns

An instance of self with simulated output of class stateprobs stored in stateprobs_.

Method sim_qalys()

Simulate quality-adjusted life-years (QALYs) as a function of stateprobs_ and utility_model. See vignette("expected-values") for details.

Usage

CohortDtstm$sim_qalys(
  dr = 0.03,
  integrate_method = c("trapz", "riemann_left", "riemann_right"),
  lys = TRUE
)

Arguments

dr

Discount rate.

integrate_method

Method used to integrate state values when computing (QALYs).

lys

If TRUE, then life-years are simulated in addition to QALYs.

Returns

An instance of self with simulated output of class qalys stored in qalys_.

Method sim_costs()

Simulate costs as a function of stateprobs_ and cost_models. See vignette("expected-values") for details.

Usage

CohortDtstm$sim_costs(
  dr = 0.03,
  integrate_method = c("trapz", "riemann_left", "riemann_right")
)

Arguments

dr

Discount rate.

integrate_method

Method used to integrate state values when computing costs.

Returns

An instance of self with simulated output of class costs stored in costs_.

Method summarize()

Summarize costs and QALYs so that cost-effectiveness analysis can be performed. See summarize_ce().

Usage

CohortDtstm$summarize(by_grp = FALSE)

Arguments

by_grp

If TRUE, then costs and QALYs are computed by subgroup. If FALSE, then costs and QALYs are aggregated across all patients (and subgroups).

Method clone()

The objects of this class are cloneable with this method.

Usage

CohortDtstm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("data.table")
library("nnet")
transitions_data <- data.table(multinom3_exdata$transitions)

# Treatment strategies, target population, and model structure
n_patients <- 4
patients <- transitions_data[year == 1, .(patient_id, age, female)][
 sort(sample.int(nrow(transitions_data[year == 1]), n_patients))][
   , grp_id := 1:n_patients]
hesim_dat <- hesim_data(
  patients = patients,
  strategies = data.table(strategy_id = 1:2,
                          strategy_name = c("Reference", "Intervention")),
  states = data.table(state_id = c(1, 2),
                      state_name = c("Healthy", "Sick")) # Non-death health states
)
tmat <- rbind(c(0, 1, 2),
              c(NA, 0, 1),
              c(NA, NA, NA))

# Parameter estimation
## Multinomial logistic regression
data_healthy <- transitions_data[state_from == "Healthy"]
fit_healthy <- multinom(state_to ~ strategy_name + female + age,
                        data = data_healthy, trace = FALSE)
data_sick <- droplevels(transitions_data[state_from == "Sick"])
fit_sick <- multinom(state_to ~ strategy_name + female + age,
                     data = data_sick, trace = FALSE)
transfits <- multinom_list(healthy = fit_healthy, sick = fit_sick)

## Utility
utility_tbl <- stateval_tbl(multinom3_exdata$utility,
                            dist = "beta")

## Costs
drugcost_tbl <- stateval_tbl(multinom3_exdata$costs$drugs,
                             dist = "fixed")
medcost_tbl <- stateval_tbl(multinom3_exdata$costs$medical,
                            dist = "gamma")

# Economic model
n_samples <- 3

## Construct model
### Transitions
transmod_data <- expand(hesim_dat)
transmod <- create_CohortDtstmTrans(transfits,
                                    input_data = transmod_data,
                                    trans_mat = tmat,
                                    n = n_samples)

### Utility
utilitymod <- create_StateVals(utility_tbl, n = n_samples, hesim_data = hesim_dat)

### Costs
drugcostmod <- create_StateVals(drugcost_tbl, n = n_samples, hesim_data = hesim_dat)
medcostmod <- create_StateVals(medcost_tbl, n = n_samples, hesim_data = hesim_dat)
costmods <- list(Drug = drugcostmod,
                 Medical = medcostmod)

### Combine
econmod <- CohortDtstm$new(trans_model = transmod,
                           utility_model = utilitymod,
                           cost_models = costmods)

## Simulate outcomes
econmod$sim_stateprobs(n_cycles = 20)
econmod$sim_qalys(dr = .03)
econmod$sim_costs(dr = .03)
econmod$summarize()

#> $costs
#>     category   dr sample strategy_id     costs grp_id
#>  1:     Drug 0.03      1           1  3807.719      1
#>  2:     Drug 0.03      1           2 10590.957      1
#>  3:     Drug 0.03      2           1  3932.434      1
#>  4:     Drug 0.03      2           2 10581.993      1
#>  5:     Drug 0.03      3           1  3911.340      1
#>  6:     Drug 0.03      3           2 10831.162      1
#>  7:  Medical 0.03      1           1  4437.702      1
#>  8:  Medical 0.03      1           2  5027.484      1
#>  9:  Medical 0.03      2           1  4544.539      1
#> 10:  Medical 0.03      2           2  5018.724      1
#> 11:  Medical 0.03      3           1  4503.276      1
#> 12:  Medical 0.03      3           2  5090.760      1
#> 13:    total 0.03      1           1  8245.422      1
#> 14:    total 0.03      1           2 15618.441      1
#> 15:    total 0.03      2           1  8476.973      1
#> 16:    total 0.03      2           2 15600.717      1
#> 17:    total 0.03      3           1  8414.616      1
#> 18:    total 0.03      3           2 15921.921      1
#> 
#> $qalys
#>      dr sample strategy_id    qalys grp_id
#> 1: 0.03      1           1 1.632688      1
#> 2: 0.03      1           2 1.831105      1
#> 3: 0.03      2           1 1.750890      1
#> 4: 0.03      2           2 1.902199      1
#> 5: 0.03      3           1 1.658167      1
#> 6: 0.03      3           2 1.858867      1
#> 
#> attr(,"class")
#> [1] "ce"

econmod$summarize(by_grp = TRUE)

#> $costs
#>     category   dr sample strategy_id grp_id     costs
#>  1:     Drug 0.03      1           1      1  4075.876
#>  2:     Drug 0.03      1           1      2  4063.226
#>  3:     Drug 0.03      1           1      3  3553.921
#>  4:     Drug 0.03      1           1      4  3537.854
#>  5:     Drug 0.03      1           2      1 11388.508
#>  6:     Drug 0.03      1           2      2 11350.835
#>  7:     Drug 0.03      1           2      3  9836.155
#>  8:     Drug 0.03      1           2      4  9788.331
#>  9:     Drug 0.03      2           1      1  4231.331
#> 10:     Drug 0.03      2           1      2  4218.784
#> 11:     Drug 0.03      2           1      3  3647.792
#> 12:     Drug 0.03      2           1      4  3631.829
#> 13:     Drug 0.03      2           2      1 11454.003
#> 14:     Drug 0.03      2           2      2 11417.569
#> 15:     Drug 0.03      2           2      3  9751.235
#> 16:     Drug 0.03      2           2      4  9705.166
#> 17:     Drug 0.03      3           1      1  4148.988
#> 18:     Drug 0.03      3           1      2  4137.455
#> 19:     Drug 0.03      3           1      3  3687.251
#> 20:     Drug 0.03      3           1      4  3671.668
#> 21:     Drug 0.03      3           2      1 11525.527
#> 22:     Drug 0.03      3           2      2 11491.492
#> 23:     Drug 0.03      3           2      3 10176.792
#> 24:     Drug 0.03      3           2      4 10130.834
#> 25:  Medical 0.03      1           1      1  4747.860
#> 26:  Medical 0.03      1           1      2  4733.250
#> 27:  Medical 0.03      1           1      3  4144.020
#> 28:  Medical 0.03      1           1      4  4125.679
#> 29:  Medical 0.03      1           2      1  5409.883
#> 30:  Medical 0.03      1           2      2  5391.863
#> 31:  Medical 0.03      1           2      3  4665.409
#> 32:  Medical 0.03      1           2      4  4642.781
#> 33:  Medical 0.03      2           1      1  4878.555
#> 34:  Medical 0.03      2           1      2  4864.507
#> 35:  Medical 0.03      2           1      3  4226.410
#> 36:  Medical 0.03      2           1      4  4208.683
#> 37:  Medical 0.03      2           2      1  5424.696
#> 38:  Medical 0.03      2           2      2  5407.727
#> 39:  Medical 0.03      2           2      3  4631.883
#> 40:  Medical 0.03      2           2      4  4610.589
#> 41:  Medical 0.03      3           1      1  4761.948
#> 42:  Medical 0.03      3           1      2  4749.153
#> 43:  Medical 0.03      3           1      3  4259.595
#> 44:  Medical 0.03      3           1      4  4242.408
#> 45:  Medical 0.03      3           2      1  5405.658
#> 46:  Medical 0.03      3           2      2  5389.967
#> 47:  Medical 0.03      3           2      3  4794.243
#> 48:  Medical 0.03      3           2      4  4773.170
#> 49:    total 0.03      1           1      1  8823.735
#> 50:    total 0.03      1           1      2  8796.476
#> 51:    total 0.03      1           1      3  7697.942
#> 52:    total 0.03      1           1      4  7663.533
#> 53:    total 0.03      1           2      1 16798.391
#> 54:    total 0.03      1           2      2 16742.698
#> 55:    total 0.03      1           2      3 14501.564
#> 56:    total 0.03      1           2      4 14431.112
#> 57:    total 0.03      2           1      1  9109.885
#> 58:    total 0.03      2           1      2  9083.291
#> 59:    total 0.03      2           1      3  7874.202
#> 60:    total 0.03      2           1      4  7840.512
#> 61:    total 0.03      2           2      1 16878.700
#> 62:    total 0.03      2           2      2 16825.296
#> 63:    total 0.03      2           2      3 14383.118
#> 64:    total 0.03      2           2      4 14315.755
#> 65:    total 0.03      3           1      1  8910.935
#> 66:    total 0.03      3           1      2  8886.608
#> 67:    total 0.03      3           1      3  7946.846
#> 68:    total 0.03      3           1      4  7914.076
#> 69:    total 0.03      3           2      1 16931.185
#> 70:    total 0.03      3           2      2 16881.459
#> 71:    total 0.03      3           2      3 14971.036
#> 72:    total 0.03      3           2      4 14904.004
#>     category   dr sample strategy_id grp_id     costs
#> 
#> $qalys
#>       dr sample strategy_id grp_id    qalys
#>  1: 0.03      1           1      1 1.747286
#>  2: 0.03      1           1      2 1.741883
#>  3: 0.03      1           1      3 1.524205
#>  4: 0.03      1           1      4 1.517378
#>  5: 0.03      1           2      1 1.969613
#>  6: 0.03      1           2      2 1.963078
#>  7: 0.03      1           2      3 1.699994
#>  8: 0.03      1           2      4 1.691737
#>  9: 0.03      2           1      1 1.882394
#> 10: 0.03      2           1      2 1.876870
#> 11: 0.03      2           1      3 1.625651
#> 12: 0.03      2           1      4 1.618644
#> 13: 0.03      2           2      1 2.057900
#> 14: 0.03      2           2      2 2.051393
#> 15: 0.03      2           2      3 1.753854
#> 16: 0.03      2           2      4 1.745650
#> 17: 0.03      3           1      1 1.755687
#> 18: 0.03      3           1      2 1.750902
#> 19: 0.03      3           1      3 1.566260
#> 20: 0.03      3           1      4 1.559817
#> 21: 0.03      3           2      1 1.975560
#> 22: 0.03      3           2      2 1.969785
#> 23: 0.03      3           2      3 1.748949
#> 24: 0.03      3           2      4 1.741175
#>       dr sample strategy_id grp_id    qalys
#> 
#> attr(,"class")
#> [1] "ce"