Transitions for a cohort discrete time state transition model

Simulate health state transitions in a cohort discrete time state transition model.

Format

An R6::R6Class object.

See also

Public fields

params

Parameters for simulating health state transitions. Currently supports objects of class tparams_transprobs.

input_data

An object of class input_mats.

trans_mat

A transition matrix describing the states and transitions in a discrete-time multi-state model. The format should be the same as in IndivCtstmTrans. Transitions that are not possible should be NA and the reference categories for each row (e.g., from a multinomial logistic regression) should be 0.

start_stateprobs

A vector with length equal to the number of health states containing the probability that the cohort is in each health state at the start of the simulation. For example, if there were three states and the cohort began the simulation in state 1, then start_stateprobs = c(1, 0, 0). If NULL, then a vector with the first element equal to 1 and all remaining elements equal to 0.

cycle_length

The length of a model cycle in terms of years. The default is 1 meaning that model cycles are 1 year long.

Methods

Public methods

CohortDtstmTrans$new()

CohortDtstmTrans$sim_stateprobs()

CohortDtstmTrans$clone()

Method new()
Create a new CohortDtstmTrans object.
Usage

CohortDtstmTrans$new(
  params,
  input_data = NULL,
  trans_mat = NULL,
  start_stateprobs = NULL,
  cycle_length = 1
)

Arguments

params

The params field.

input_data

The input_data field.

trans_mat

The trans_mat field.

start_stateprobs

The start_stateprobs field.

cycle_length

The cycle_length field.

Returns
A new CohortDtstmTrans object.

Method sim_stateprobs()
Simulate probability of being in each health state during each model cycle.
Usage

CohortDtstmTrans$sim_stateprobs(n_cycles)

Arguments

n_cycles

The number of model cycles to simulate the model for.

Returns
An object of class stateprobs.

Method clone()
The objects of this class are cloneable with this method.
Usage

CohortDtstmTrans$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.