An R6 base class for continuous time state transition models

Contains methods that can be used to summarize both individual- and cohort-level continuous time state transition models. That is, this class is relevant for both Markov and semi-Markov multi-state models and does not depend on the methodology used for prediction of state probabilities.

Format

An R6::R6Class object.

See also

Methods

Public methods

CtstmTrans$hazard()

CtstmTrans$cumhazard()

CtstmTrans$clone()

Method hazard()

Predict the hazard functions for each health state transition.

Usage

CtstmTrans$hazard(t)

Arguments

t

A numeric vector of times.

Returns

A data.table with columns trans (the transition number), sample, strategy_id, grp_id, t, and hazard.

Method cumhazard()

Predict the cumulative hazard functions for each health state transition.

Usage

CtstmTrans$cumhazard(t)

Arguments

t

A numeric vector of times.

Returns

A data.table with columns trans, sample, strategy_id, grp_id, t, and hazard.

Method clone()

The objects of this class are cloneable with this method.

Usage

CtstmTrans$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.