Individual-level continuous time state transition model

Simulate outcomes from an individual-level continuous time state transition model (CTSTM). The class is currently only valid for semi-Markov (i.e., "clock-reset") models; however, functionality will soon be added for Markov (i.e., "clock-forward") models and mixtures of clock-forward and clock-reset models.

IndivCtstm

Format

R6Class object.

Fields

trans_model

Model for health state transitions. Must be an object of class IndivCtstmTrans.

utility_model

The model used to predict utility by health state. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

disprog_

A data.table simulated using sim_disease containing the following columns:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

The patient ID.

from

The health state ID transitioned from.

to

The health state ID transitioned to.

final

An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time_start

The time at the start of the interval.

time_stop

The time at the end of the interval.

stateprobs_

A data.table of health state probabilities as a function of time simulated using sim_stateprobs. See the description of the output of sim_stateprobs in IndivCtstmTrans.

qalys_

A data.table of quality-adjusted life-years (QALYs) simulated using sim_qalys. Columns are:

sample

A random sample from the probabilistic sensitivity analysis (PSA).

strategy_id

The treatment strategy ID.

patient_id

A patient ID.

dr

The discount rate.

qalys

Simulated QALYs.

lys

Simulated life-years.

Note that the lys column is only computed if lys = TRUE.

costs_

A data.table of costs by category simulated using sim_costs. Columns are:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

A patient ID.

dr

The discount rate.

category

The cost category.

costs

Simulated costs.

Methods

new(trans_model = NULL, utility_model = NULL, cost_models = NULL)

Constructor for the class.

sim_disease(max_t = 100, max_age = 100)

Simulate disease progression.

max_t: A scalar or vector denoting the number of time periods to simulate the model. If a vector, must be equal to the number of simulated patients.

max_age: A scalar or vector denoting the maximum age to simulate each patient until. If a vector, must be equal to the number of simulated patients.

Returns an instance of self with simulated output stored in disprog_.

sim_stateprobs(t)

Simulate the probability of being in each health state as a function of time using the simulation output stored in disprog_.

t: A numeric vector of times.

Returns an instance of self with simulated output stored in stateprobs_.

sim_qalys(dr = .03, type = c("predict", "random"), by_patient = FALSE, lys = TRUE)

Compute simulated (mean discounted) QALYs using the simulation output stored in disprog_ by random sample from the PSA, treatment strategy, health state, and (optionally) patient.

dr: Discount rate to apply to QALYs. May be a vector in which case QALYs are calculated for each element in dr.

type: predict for mean values or random for random samples as in $sim() in StateVals.

by_patient: If TRUE, QALYs are computed at the patient level. If FALSE, QALYs are averaged across patients by health state.

lys: If TRUE, then life-years are computed in addition to QALYs.

Returns an instance of self with simulated output stored in qalys_.

sim_costs(dr = .03, type = c("predict", "random"), by_patient = FALSE, max_t = Inf)

Compute simulated (mean discounted) costs using the simulation output stored in disprog_ by random sample from the PSA, treatment strategy, health state, and (optionally) patient.

dr: Discount rate to apply to costs. May be a vector in which case costs are calculated for each element in dr.

type: predict for mean values or random for random samples as in $sim() in StateVals.

by_patient: If TRUE, costs are computed at the patient level. If FALSE, costs are averaged across patients by health state.

max_t: Maximum time duration to compute costs once a patient has entered a (new) health state. By default, equal to Inf, so that costs are computed over the entire duration that a patient is in a given health state. If time varies by each cost category, then time can also be passed as a numeric vector of length equal to the number of cost categories (e.g., c(1, 2, Inf, 3) for a model with 4 cost categories).

Returns an instance of self with simulated output stored in costs_.

,
summarize()

Produce a summary measure of costs and QALYs by random sample from the PSA and treatment strategy. Returns an object of summary cost-effectiveness output of class ce.

See also

Examples

library("flexsurv")

# Treatment strategies,target population, and health states
dt_strategies <- data.frame(strategy_id = c(1, 2))
dt_patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))
dt_states <- data.frame(state_id = c(1, 2))

# Parameter estimation
n_samples <- 2

## Multi-state model
fits <- vector(length = 3, mode = "list")
surv_dat <- data.frame(ctstm3_exdata$transitions)
for (i in 1:length(fits)){
  fits[[i]] <- flexsurvreg(Surv(years, status) ~ factor(strategy_id),
                           data = surv_dat,
                           subset = (trans == i),
                           dist = "weibull")
}
fits <- flexsurvreg_list(fits)
tmat <- rbind(c(NA, 1, 2),
              c(NA, NA, 3),
              c(NA, NA, NA))

## Utility
beta_params <- mom_beta(ctstm3_exdata$utility$mean, ctstm3_exdata$utility$se)
utility_posterior <- matrix(rbeta(n_samples * 2, shape1 = beta_params$shape1,
                                  shape2 = beta_params$shape2),
                            nrow = n_samples, byrow = TRUE)
## Costs
drugcosts <- ctstm3_exdata$costs$drugs$costs

# Economic model
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients,
                        states = dt_states)

## Create transition model (i.e., multi-state model)
transmod_data <- expand(hesim_dat)
transmod <- create_IndivCtstmTrans(fits, data = transmod_data, trans_mat = tmat,
                                   n = n_samples)

## Create utility model
utility_means <- stateval_means(values = utility_posterior,
                                strategy_id = dt_strategies$strategy_id,
                                patient_id = dt_patients$patient_id)
utilitymod <- create_StateVals(utility_means)

## Create (drug) cost model
drugcost_array <- array(NA, dim = c(n_samples, 2, 2))
drugcost_array[, , 1] <- drugcosts[1]
drugcost_array[, , 2] <- drugcosts[2]
drugcost_means <- stateval_means(values = drugcost_array,
                                strategy_id = dt_strategies$strategy_id,
                                patient_id = dt_patients$patient_id)

drugcostmod <- create_StateVals(drugcost_means)

## Simulate outcomes
ictstm <- IndivCtstm$new(trans_model = transmod,
                         utility_model = utilitymod,
                         cost_models = list(drugs = drugcostmod))
head(ictstm$sim_disease()$disprog_$sim)

#> NULL

head(ictstm$sim_stateprobs(t = c(0, 5, 10))$stateprobs_[t == 5])

#>    sample strategy_id state_id t      prob
#> 1:      1           1        1 5 0.0000000
#> 2:      1           1        2 5 0.0000000
#> 3:      1           1        3 5 1.0000000
#> 4:      1           2        1 5 0.3333333
#> 5:      1           2        2 5 0.0000000
#> 6:      1           2        3 5 0.6666667

ictstm$sim_qalys(dr = .03)
ictstm$sim_costs(dr = .03)
head(ictstm$summarize())

#> $costs
#>    category   dr sample strategy_id     costs
#> 1:    drugs 0.03      1           1  4139.011
#> 2:    drugs 0.03      1           2 51773.575
#> 3:    drugs 0.03      2           1 40870.709
#> 4:    drugs 0.03      2           2 33812.944
#> 5:    total 0.03      1           1  4139.011
#> 6:    total 0.03      1           2 51773.575
#> 7:    total 0.03      2           1 40870.709
#> 8:    total 0.03      2           2 33812.944
#> 
#> $qalys
#>      dr sample strategy_id    qalys
#> 1: 0.03      1           1 0.733864
#> 2: 0.03      1           2 4.589836
#> 3: 0.03      2           1 7.862465
#> 4: 0.03      2           2 3.248173
#>