Individual-level continuous time state transition model

Simulate outcomes from an individual-level continuous time state transition model (CTSTM). The class is only valid for semi-Markov (i.e., "clock-reset") models.

IndivCtstm

Format

R6Class object.

Fields

trans_model

Model for health state transitions. Must be an object of class CtstmTrans.

utility_model

The model used to predict utility by health state. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

disease_prog_

Disease progression simulated using sim_disease; that is, a tidy data.table containing simulated trajectories through the multi-state model. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

The patient ID.

state_id

The health state ID.

final

An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time

The time when a patient enters a new health state.

stateprobs_

A tidy data.table of health state probabilities as a function of time simulated using sim_stateprobs. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

state_id

The health state ID.

t

The time at which a state probability is computed.

prob

The probability of being in a given health state.

Methods

new(trans_model, utility_model = NULL, cost_models = NULL)

Constructor for the class.

sim_disease(max_t = 100, max_age = 100)

Simulate disease progression.

max_t: Maximum time to simulate disease until.

max_age: Maximum age that a patient can live. Upon reaching this age the simulation stops and a patient moves to the "death state" specified by death_state.

The simulated output is stored in disease_prog_.

sim_stateprobs(t)

Simulate the probability of being in each health state as a function of time.

t: A numeric vector of times.

The simulated output is stored in stateprobs_.

See also

create_CtstmTrans, CtstmTrans

Examples

library("flexsurv")
library("data.table")

# Simulation data
dt_strategies <- data.frame(strategy_id = c(1, 2, 3))
dt_patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))

# Multi-state model
fits <- vector(length = 3, mode = "list")
dat <- data.table(bosms3)
for (i in 1:length(fits)){
  fits[[i]] <- flexsurvreg(Surv(years, status) ~ 1, data = dat[trans == i],
                           dist = "exp")
}
fits <- flexsurvreg_list(fits)
tmat <- rbind(c(NA, 1, 2),
              c(NA, NA, 3),
              c(NA, NA, NA))

# Simulation model
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients)
fits_data <- expand_hesim_data(hesim_dat)
transmod <- create_CtstmTrans(fits, data = fits_data, trans_mat = tmat,
                              point_estimate = TRUE)
ictstm <- IndivCtstm$new(trans_model = transmod)
head(ictstm$sim_disease()$disease_prog_)

#>    sample strategy_id patient_id from to final time_start time_stop
#> 1:      1           1          1    1  2     0   0.000000 12.490048
#> 2:      1           1          1    2  3     1  12.490048 13.253595
#> 3:      1           1          2    1  3     1   0.000000  8.658507
#> 4:      1           1          3    1  2     0   0.000000  4.488498
#> 5:      1           1          3    2  3     1   4.488498  6.553614
#> 6:      1           2          1    1  2     0   0.000000  0.125005

head(ictstm$sim_stateprobs(t = c(0, 5, 10))$stateprobs_)

#>    sample strategy_id state_id  t      prob
#> 1:      1           1        1  0 1.0000000
#> 2:      1           1        1  5 0.6666667
#> 3:      1           1        1 10 0.3333333
#> 4:      1           1        2  0 0.0000000
#> 5:      1           1        2  5 0.3333333
#> 6:      1           1        2 10 0.0000000