Individual-level continuous time state transition model

Simulate outcomes from an individual-level continuous time state transition model (CTSTM). The class is currently only valid for semi-Markov (i.e., "clock-reset") models; however, functionality will soon be added for Markov (i.e., "clock-forward") models.

IndivCtstm

Format

R6Class object.

Fields

trans_model

Model for health state transitions. Must be an object of class CtstmTrans.

utility_model

The model used to predict utility by health state. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

disease_prog_

Disease progression simulated using sim_disease; that is, a tidy data.table containing simulated trajectories through the multi-state model. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

The patient ID.

from

The health state ID transitioned from.

to

The health state ID transitioned to.

final

An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time_start

The time at the start of the interval.

time_start

The time at the end of the interval.

stateprobs_

A tidy data.table of health state probabilities as a function of time simulated using sim_stateprobs. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

state_id

The health state ID.

t

The time at which a state probability is computed.

prob

The probability of being in a given health state.

qalys_

A tidy data.table of quality-adjusted life-years (QALYs) simulated using sim_qalys. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

A patient ID.

dr

The discount rate.

qalys

Simulated QALYs.

costs_

A tidy data.table of costs by category simulated using sim_costs. Columns include:

sample

A random sample from the PSA.

strategy_id

The treatment strategy ID.

patient_id

A patient ID.

dr

The discount rate.

category

The cost category.

costs

Simulated costs.

Methods

new(trans_model, utility_model = NULL, cost_models = NULL)

Constructor for the class.

sim_disease(max_t = 100, max_age = 100)

Simulate disease progression.

max_t: Maximum time to simulate disease until.

max_age: Maximum age that a patient can live. Upon reaching this age the simulation stops and a patient moves to the "death state" specified by death_state.

The simulated output is stored in disease_prog_.

sim_stateprobs(t)

Simulate the probability of being in each health state as a function of time.

t: A numeric vector of times.

The simulated output is stored in stateprobs_.

sim_qalys(dr = .03, type = c("predict", "random")

Compute simulated (discounted) QALYs using the simulation output generated by sim_disease and stored in disease_prog_ by patient and treatment strategy.

dr: Discount rate to apply to QALYs. May be a vector in which case QALYs are calculated for each element in dr.

type: predict for mean values or random for random samples as in $sim() in StateVals.

The simulated output is stored in qalys_.

sim_costs(dr = .03, type = c("predict", "random")

Compute simulated (discounted) costs using the simulation output generated by sim_disease and stored in disease_prog_ by patient and treatment strategy.

dr: Discount rate to apply to costs. May be a vector in which case costs are calculated for each element in dr.

type: predict for mean values or random for random samples as in $sim() in StateVals.

The simulated output is stored in costs_.

,
summarize(stat = mean)

Summarize costs and QALYs to produce a summary measure of costs and health outcomes across all patients by treatment strategy.

stat: The statistic used to summarize costs and QALYs. The default is the mean but any R function can be passed to stat.

Returns an object of summary cost-effectiveness output of class ce.

See also

Examples

library("flexsurv")

# Treatment strategies,target population, and health states
dt_strategies <- data.frame(strategy_id = c(1, 2, 3))
dt_patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))
dt_states <- data.frame(state_id = c(1, 2))

# Parameter estimation
n_samples <- 3

## Multi-state model
fits <- vector(length = 3, mode = "list")
surv_dat <- data.frame(ctstm3_exdata$transitions)
for (i in 1:length(fits)){
  fits[[i]] <- flexsurvreg(Surv(years, status) ~ strategy_id,
                           data = surv_dat,
                           subset = (trans == i),
                           dist = "weibull")
}
fits <- flexsurvreg_list(fits)
tmat <- rbind(c(NA, 1, 2),
              c(NA, NA, 3),
              c(NA, NA, NA))

## Utility
beta_params <- mom_beta(ctstm3_exdata$utility$mean, ctstm3_exdata$utility$se)
utility1 <- rbeta(n_samples, shape1 = beta_params$shape1[1],
                  shape2 = beta_params$shape2[1])
utility2 <- rbeta(n_samples, shape1 = beta_params$shape1[2],
                  shape2 = beta_params$shape2[2])
utility_params <- params_lm(coefs  = cbind(utility1, utility2))

# Costs
drugcost_params <- params_lm(matrix(rep(ctstm3_exdata$costs$drugs$costs, n_samples),
                             nrow = n_samples, byrow = TRUE))

# Decision model
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients,
                        states = dt_states)

## Create transition model (i.e., multi-state model)
transmod_data <- expand_hesim_data(hesim_dat)
transmod <- create_CtstmTrans(fits, data = transmod_data, trans_mat = tmat,
                              n = n_samples)

## Create costs and utility models
statevals_edata <- expand_hesim_data(hesim_dat, by = c("strategies", "patients",
                                                       "states"))
input_dat <- create_input_data(formula_list(mu = ~ -1 + factor(state_id)),
                               data = statevals_edata)
utilmod <- StateVals$new(data = input_dat, params = utility_params)
drugcostmod <- StateVals$new(data = input_dat, params = drugcost_params)

## Simulate outcomes
ictstm <- IndivCtstm$new(trans_model = transmod,
                         utility_model = utilmod,
                         cost_models = list(drugs = drugcostmod))
head(ictstm$sim_disease()$disease_prog_)

#>    sample strategy_id patient_id from to final time_start  time_stop
#> 1:      1           1          1    1  3     1  0.0000000 0.51147464
#> 2:      1           1          2    1  2     0  0.0000000 0.35247378
#> 3:      1           1          2    2  3     1  0.3524738 0.57498942
#> 4:      1           1          3    1  3     1  0.0000000 0.00344779
#> 5:      1           2          1    1  3     1  0.0000000 3.57722655
#> 6:      1           2          2    1  2     0  0.0000000 2.48987431

head(ictstm$sim_stateprobs(t = c(0, 5, 10))$stateprobs_[t == 5])

#>    sample strategy_id state_id t prob
#> 1:      1           1        1 5    0
#> 2:      1           1        2 5    0
#> 3:      1           1        3 5    1
#> 4:      1           2        1 5    0
#> 5:      1           2        2 5    0
#> 6:      1           2        3 5    1

ictstm$sim_qalys(dr = .03)
ictstm$sim_costs(dr = .03)
head(ictstm$summarize())

#> $costs
#>    category   dr sample strategy_id     costs
#> 1:    drugs 0.03      1           1  2167.531
#> 2:    drugs 0.03      1           2 14527.999
#> 3:    drugs 0.03      1           3 35689.335
#> 4:    drugs 0.03      2           1 18704.018
#> 5:    drugs 0.03      2           2  6254.678
#> 6:    drugs 0.03      2           3 14099.266
#> 7:    drugs 0.03      3           1 17229.288
#> 8:    drugs 0.03      3           2 17071.675
#> 9:    drugs 0.03      3           3 23553.444
#> 
#> $qalys
#>      dr sample strategy_id     qalys
#> 1: 0.03      1           1 0.2299682
#> 2: 0.03      1           2 1.8480889
#> 3: 0.03      1           3 4.0020934
#> 4: 0.03      2           1 1.8321787
#> 5: 0.03      2           2 0.4957363
#> 6: 0.03      2           3 0.9023704
#> 7: 0.03      3           1 1.6510278
#> 8: 0.03      3           2 1.6232142
#> 9: 0.03      3           3 2.1365384
#>