Individual-level continuous time state transition model

Simulate outcomes from an individual-level continuous time state transition model (CTSTM) from a fitted multi-state model. The class supports "clock-reset" (i.e., semi-Markov), "clock-forward" (i.e., Markov), and mixtures of clock-reset and clock-forward models as described in IndivCtstmTrans.

Format

An R6::R6Class object.

See also

Public fields

trans_model

The model for health state transitions. Must be an object of class IndivCtstmTrans.

utility_model

The model for health state utility. Must be an object of class StateVals.

cost_models

The models used to predict costs by health state. Must be a list of objects of class StateVals, where each element of the list represents a different cost category.

disprog_

A data.table simulated using $sim_disease() containing the following columns:

sample: A random sample from the PSA.

strategy_id: The treatment strategy ID.

patient_id: The patient ID.

from: The health state ID transitioned from.

to: The health state ID transitioned to.

final: An indicator equal to 1 if a patient is in their final health state during the simulation and 0 otherwise.

time_start: The time at the start of the interval.

time_stop: The time at the end of the interval.

stateprobs_

An object of class stateprobs simulated using $sim_stateprobs().

qalys_

An object of class qalys simulated using $sim_qalys().

costs_

An object of class costs simulated using $sim_costs().

Methods

Public methods

IndivCtstm$new()

IndivCtstm$sim_disease()

IndivCtstm$sim_stateprobs()

IndivCtstm$sim_qalys()

IndivCtstm$sim_costs()

IndivCtstm$summarize()

IndivCtstm$clone()

Method new()

Create a new IndivCtstm object.

Usage

IndivCtstm$new(trans_model = NULL, utility_model = NULL, cost_models = NULL)

Arguments

trans_model

The trans_model field.

utility_model

The utility_model field.

cost_models

The cost_models field.

Returns

A new IndivCtstm object.

Method sim_disease()

Simulate disease progression (i.e., individual trajectories through a multi-state model using an individual patient simulation).

Usage

IndivCtstm$sim_disease(max_t = 100, max_age = 100, progress = NULL)

Arguments

max_t

A scalar or vector denoting the length of time to simulate the model. If a vector, must be equal to the number of simulated patients.

max_age

A scalar or vector denoting the maximum age to simulate each patient until. If a vector, must be equal to the number of simulated patients.

progress

An integer, specifying the PSA iteration (i.e., sample) that should be printed every progress PSA iterations. For example, if progress = 2, then every second PSA iteration is printed. Default is NULL, in which case no output is printed.

Returns

An instance of self with simulated output stored in disprog_.

Method sim_stateprobs()

Simulate health state probabilities as a function of time using the simulation output stored in disprog.

Usage

IndivCtstm$sim_stateprobs(t)

Arguments

t

TA numeric vector of times.

Returns

An instance of self with simulated output of class stateprobs stored in stateprobs_.

Method sim_qalys()

Simulate quality-adjusted life-years (QALYs) as a function of disprog_ and utility_model. See vignette("expected-values") for details.

Usage

IndivCtstm$sim_qalys(
  dr = 0.03,
  type = c("predict", "random"),
  lys = TRUE,
  by_patient = FALSE
)

Arguments

dr

Discount rate.

type

"predict" for mean values or "random" for random samples as in $sim() in StateVals.

lys

If TRUE, then life-years are simulated in addition to QALYs.

by_patient

If TRUE, then QALYs are computed at the patient level. If FALSE, then QALYs are averaged across patients by health state.

Returns

An instance of self with simulated output of class qalys stored in qalys_.

Method sim_costs()

Simulate costs as a function of disprog_ and cost_models. See vignette("expected-values") for details.

Usage

IndivCtstm$sim_costs(
  dr = 0.03,
  type = c("predict", "random"),
  by_patient = FALSE,
  max_t = Inf
)

Arguments

dr

Discount rate.

type

"predict" for mean values or "random" for random samples as in $sim() in StateVals.

by_patient

If TRUE, then QALYs are computed at the patient level. If FALSE, then QALYs are averaged across patients by health state.

max_t

Maximum time duration to compute costs once a patient has entered a (new) health state. By default, equal to Inf, so that costs are computed over the entire duration that a patient is in a given health state. If time varies by each cost category, then time can also be passed as a numeric vector of length equal to the number of cost categories (e.g., c(1, 2, Inf, 3) for a model with four cost categories).

Returns

An instance of self with simulated output of class costs stored in costs_.

Method summarize()

Summarize costs and QALYs so that cost-effectiveness analysis can be performed. See summarize_ce().

Usage

IndivCtstm$summarize(by_grp = FALSE)

Arguments

by_grp

If TRUE, then costs and QALYs are computed by subgroup. If FALSE, then costs and QALYs are aggregated across all patients (and subgroups).

Method clone()

The objects of this class are cloneable with this method.

Usage

IndivCtstm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("flexsurv")

#> Loading required package: survival

# Treatment strategies, target population, and model structure
strategies <- data.frame(strategy_id = c(1, 2))
patients <- data.frame(patient_id = seq(1, 3),
                       age = c(45, 50, 60),
                       female = c(0, 0, 1))
states <- data.frame(state_id = c(1, 2))
hesim_dat <- hesim_data(strategies = strategies,
                        patients = patients,
                        states = states)

# Parameter estimation
## Multi-state model
tmat <- rbind(c(NA, 1, 2),
              c(3, NA, 4),
              c(NA, NA, NA))
fits <- vector(length = max(tmat, na.rm = TRUE), mode = "list")
surv_dat <- data.frame(mstate3_exdata$transitions)
for (i in 1:length(fits)){
  fits[[i]] <- flexsurvreg(Surv(years, status) ~ factor(strategy_id),
                           data = surv_dat,
                           subset = (trans == i),
                           dist = "weibull")
}
fits <- flexsurvreg_list(fits)

## Utility
utility_tbl <- stateval_tbl(data.frame(state_id = states$state_id,
                                       mean = mstate3_exdata$utility$mean,
                                       se = mstate3_exdata$utility$se),
                            dist = "beta",
                            hesim_data = hesim_dat)
## Costs
drugcost_tbl <- stateval_tbl(data.frame(strategy_id = strategies$strategy_id,
                                        est = mstate3_exdata$costs$drugs$costs),
                             dist = "fixed",
                             hesim_data = hesim_dat)
medcost_tbl <- stateval_tbl(data.frame(state_id = states$state_id,
                                       mean = mstate3_exdata$costs$medical$mean,
                                       se = mstate3_exdata$costs$medical$se),
                            dist = "gamma",
                            hesim_data = hesim_dat)

# Economic model
n_samples = 2

## Construct model
### Transitions 
transmod_data <- expand(hesim_dat)
transmod <- create_IndivCtstmTrans(fits, input_data = transmod_data,
                                   trans_mat = tmat,
                                   n = n_samples)

### Utility 
utilitymod <- create_StateVals(utility_tbl, n = n_samples)

### Costs
drugcostmod <- create_StateVals(drugcost_tbl, n = n_samples)
medcostmod <- create_StateVals(medcost_tbl, n = n_samples)
costmods <- list(drugs = drugcostmod,
                 medical = medcostmod)

### Combine
ictstm <- IndivCtstm$new(trans_model = transmod,
                         utility_model = utilitymod,
                         cost_models = costmods)

## Simulate outcomes
head(ictstm$sim_disease()$disprog_)

#>    sample strategy_id patient_id grp_id from to final time_start time_stop
#> 1:      1           1          1      1    1  3     1  0.0000000 0.4036627
#> 2:      1           1          2      1    1  2     0  0.0000000 5.5078506
#> 3:      1           1          2      1    2  1     0  5.5078506 5.8230571
#> 4:      1           1          2      1    1  3     1  5.8230571 6.6985224
#> 5:      1           1          3      1    1  2     0  0.0000000 0.7563655
#> 6:      1           1          3      1    2  1     0  0.7563655 0.7578541

head(ictstm$sim_stateprobs(t = c(0, 5, 10))$stateprobs_[t == 5])

#>    sample strategy_id grp_id state_id t      prob
#> 1:      1           1      1        1 5 0.3333333
#> 2:      1           1      1        2 5 0.3333333
#> 3:      1           1      1        3 5 0.3333333
#> 4:      1           2      1        1 5 0.6666667
#> 5:      1           2      1        2 5 0.0000000
#> 6:      1           2      1        3 5 0.3333333

ictstm$sim_qalys(dr = .03)
ictstm$sim_costs(dr = .03)
head(ictstm$summarize())

#> $costs
#>     category   dr sample strategy_id     costs grp_id
#>  1:    drugs 0.03      1           1 42596.065      1
#>  2:    drugs 0.03      1           2 44645.168      1
#>  3:    drugs 0.03      2           1 25228.525      1
#>  4:    drugs 0.03      2           2 89673.803      1
#>  5:  medical 0.03      1           1 10533.655      1
#>  6:  medical 0.03      1           2  4610.220      1
#>  7:  medical 0.03      2           1  5296.932      1
#>  8:  medical 0.03      2           2  9632.893      1
#>  9:    total 0.03      1           1 53129.720      1
#> 10:    total 0.03      1           2 49255.388      1
#> 11:    total 0.03      2           1 30525.457      1
#> 12:    total 0.03      2           2 99306.697      1
#> 
#> $qalys
#>      dr sample strategy_id    qalys grp_id
#> 1: 0.03      1           1 3.572185      1
#> 2: 0.03      1           2 2.606998      1
#> 3: 0.03      2           1 1.440414      1
#> 4: 0.03      2           2 2.703469      1
#>