Partitioned survival curves

Summarize n-1 survival curves for an N state partitioned survival model.

Format

An R6::R6Class object.

See also

Public fields

params

An object of class params_surv_list.

input_data

An object of class input_mats. Each row in X must be a unique treatment strategy and patient.

Methods

Public methods

PsmCurves$new()

PsmCurves$hazard()

PsmCurves$cumhazard()

PsmCurves$survival()

PsmCurves$rmst()

PsmCurves$quantile()

PsmCurves$check()

PsmCurves$clone()

Method new()

Create a new PsmCurves object.

Usage

PsmCurves$new(params, input_data)

Arguments

params

The params field.

input_data

The input_data field.

Returns

A new PsmCurves object.

Method hazard()

Predict the hazard function for each survival curve as a function of time.

Usage

PsmCurves$hazard(t)

Arguments

t

A numeric vector of times.

Returns

A data.table with columns sample, strategy_id, patient_id, grp_id, curve (the curve number), t, and hazard.

Method cumhazard()

Predict the cumulative hazard function for each survival curve as a function of time.

Usage

PsmCurves$cumhazard(t)

Arguments

t

A numeric vector of times.

Returns

A data.table with columns sample, strategy_id, patient_id, grp_id, curve, t, and cumhazard.

Method survival()

Predict the cumulative hazard function for each survival curve as a function of time.

Usage

PsmCurves$survival(t)

Arguments

t

A numeric vector of times.

Returns

A data.table with columns sample, strategy_id, patient_id, grp_id, curve, t, and survival.

Method rmst()

Predict the restricted mean survival time up until time points t for each survival curve.

Usage

PsmCurves$rmst(t, dr = 0)

Arguments

t

A numeric vector of times.

dr

Discount rate.

Returns

A data.table with columns sample, strategy_id, patient_id, grp_id, curve, t, and rmst.

Method quantile()

Predict quantiles of the survival distribution for each survival curve.

Usage

PsmCurves$quantile(p)

Arguments

p

A numeric vector of probabilities for computing quantiles.

Returns

A data.table with columns sample, strategy_id, patient_id, grp_id, curve, p and quantile.

Method check()

Input validation for class. Checks that fields are the correct type.

Usage

PsmCurves$check()

Method clone()

The objects of this class are cloneable with this method.

Usage

PsmCurves$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library("flexsurv")

# Simulation data
dt_strategies <- data.frame(strategy_id = c(1, 2, 3))
dt_patients <- data.frame(patient_id = seq(1, 3),
                          age = c(45, 50, 60),
                          female = c(0, 0, 1))
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients)

# Fit survival models
surv_est_data <- psm4_exdata$survival
fit1 <- flexsurv::flexsurvreg(Surv(endpoint1_time, endpoint1_status) ~ age,
                              data = surv_est_data, dist = "exp")
fit2 <- flexsurv::flexsurvreg(Surv(endpoint2_time, endpoint2_status) ~ age,
                              data = surv_est_data, dist = "exp")
fit3 <- flexsurv::flexsurvreg(Surv(endpoint3_time, endpoint3_status) ~ age,
                              data = surv_est_data, dist = "exp")
fits <- flexsurvreg_list(fit1, fit2, fit3)

# Form PsmCurves
surv_input_data <- expand(hesim_dat, by = c("strategies", "patients"))
psm_curves <- create_PsmCurves(fits, input_data = surv_input_data, n = 3,
                               bootstrap = TRUE, est_data = surv_est_data)

# Summarize survival curves
head(psm_curves$quantile(p = c(.25, .5, .75)))

#>    sample strategy_id patient_id grp_id curve    p  quantile
#> 1:      1           1          1      1     1 0.25 0.1773808
#> 2:      1           1          1      1     1 0.50 0.4273851
#> 3:      1           1          1      1     1 0.75 0.8547701
#> 4:      1           1          1      1     2 0.25 0.3608403
#> 5:      1           1          1      1     2 0.50 0.8694162
#> 6:      1           1          1      1     2 0.75 1.7388324

head(psm_curves$survival(t = seq(0, 3, by = .1)))

#>    sample strategy_id patient_id grp_id curve   t  survival
#> 1:      1           1          1      1     1 0.0 1.0000000
#> 2:      1           1          1      1     1 0.1 0.8502853
#> 3:      1           1          1      1     1 0.2 0.7229852
#> 4:      1           1          1      1     1 0.3 0.6147437
#> 5:      1           1          1      1     1 0.4 0.5227075
#> 6:      1           1          1      1     1 0.5 0.4444506

head(psm_curves$rmst(t = c(2, 5)))

#>    sample strategy_id patient_id grp_id curve t      rmst
#> 1:      1           1          1      1     1 2 0.5925268
#> 2:      1           1          1      1     1 5 0.6164009
#> 3:      1           1          1      1     2 2 0.9996695
#> 4:      1           1          1      1     2 5 1.2310117
#> 5:      1           1          1      1     3 2 1.2077934
#> 6:      1           1          1      1     3 5 1.6889371