Coerce to data.table

Creates a data.table that combines the transition probability matrices and ID columns from a tparams_transprobs object. This is often useful for debugging.

# S3 method for tparams_transprobs
as.data.table(x, ..., prefix = "prob_", sep = "_")

Arguments

x
A tparams_transprobs object.

...
Currently unused.

prefix, sep
Arguments passed to tpmatrix_names() for naming the transition probability columns. The states argument is based on the column names (i.e., names of the second dimension) of the $value element of x; if NULL, then states are named s1, ..., sh where h is the number of states.

Value

A data.table with one row for each transition probability matrix.

See also

Examples

# Create tparams_transprobs object
hesim_dat <- hesim_data(strategies = data.frame(strategy_id = 1:2),
                        patients = data.frame(patient_id = 1:3))
input_data <- expand(hesim_dat, by = c("strategies", "patients"))    
tpmat_id <- tpmatrix_id(input_data, n_samples = 2)      
p_12 <- runif(nrow(tpmat_id), .6, .7) + 
  .05 * (tpmat_id$strategy_id == 2)
tpmat <- tpmatrix(
  C, p_12,
  0, 1
)
tprobs <- tparams_transprobs(tpmat, tpmat_id)

# Convert to data.table
as.data.table(tprobs)

#>     sample strategy_id patient_id time_id time_start time_stop prob_s1_s1
#>  1:      1           1          1       1          0       Inf  0.3919250
#>  2:      1           1          2       1          0       Inf  0.3165667
#>  3:      1           1          3       1          0       Inf  0.3399239
#>  4:      1           2          1       1          0       Inf  0.3342792
#>  5:      1           2          2       1          0       Inf  0.3492601
#>  6:      1           2          3       1          0       Inf  0.3033607
#>  7:      2           1          1       1          0       Inf  0.3502223
#>  8:      2           1          2       1          0       Inf  0.3710233
#>  9:      2           1          3       1          0       Inf  0.3267118
#> 10:      2           2          1       1          0       Inf  0.2727478
#> 11:      2           2          2       1          0       Inf  0.2625399
#> 12:      2           2          3       1          0       Inf  0.3325059
#>     prob_s1_s2 prob_s2_s1 prob_s2_s2
#>  1:  0.6080750          0          1
#>  2:  0.6834333          0          1
#>  3:  0.6600761          0          1
#>  4:  0.6657208          0          1
#>  5:  0.6507399          0          1
#>  6:  0.6966393          0          1
#>  7:  0.6497777          0          1
#>  8:  0.6289767          0          1
#>  9:  0.6732882          0          1
#> 10:  0.7272522          0          1
#> 11:  0.7374601          0          1
#> 12:  0.6674941          0          1

as.data.table(tprobs, prefix = "")

#>     sample strategy_id patient_id time_id time_start time_stop     s1_s1
#>  1:      1           1          1       1          0       Inf 0.3919250
#>  2:      1           1          2       1          0       Inf 0.3165667
#>  3:      1           1          3       1          0       Inf 0.3399239
#>  4:      1           2          1       1          0       Inf 0.3342792
#>  5:      1           2          2       1          0       Inf 0.3492601
#>  6:      1           2          3       1          0       Inf 0.3033607
#>  7:      2           1          1       1          0       Inf 0.3502223
#>  8:      2           1          2       1          0       Inf 0.3710233
#>  9:      2           1          3       1          0       Inf 0.3267118
#> 10:      2           2          1       1          0       Inf 0.2727478
#> 11:      2           2          2       1          0       Inf 0.2625399
#> 12:      2           2          3       1          0       Inf 0.3325059
#>         s1_s2 s2_s1 s2_s2
#>  1: 0.6080750     0     1
#>  2: 0.6834333     0     1
#>  3: 0.6600761     0     1
#>  4: 0.6657208     0     1
#>  5: 0.6507399     0     1
#>  6: 0.6966393     0     1
#>  7: 0.6497777     0     1
#>  8: 0.6289767     0     1
#>  9: 0.6732882     0     1
#> 10: 0.7272522     0     1
#> 11: 0.7374601     0     1
#> 12: 0.6674941     0     1

as.data.table(tprobs, prefix = "", sep = ".")

#>     sample strategy_id patient_id time_id time_start time_stop     s1.s1
#>  1:      1           1          1       1          0       Inf 0.3919250
#>  2:      1           1          2       1          0       Inf 0.3165667
#>  3:      1           1          3       1          0       Inf 0.3399239
#>  4:      1           2          1       1          0       Inf 0.3342792
#>  5:      1           2          2       1          0       Inf 0.3492601
#>  6:      1           2          3       1          0       Inf 0.3033607
#>  7:      2           1          1       1          0       Inf 0.3502223
#>  8:      2           1          2       1          0       Inf 0.3710233
#>  9:      2           1          3       1          0       Inf 0.3267118
#> 10:      2           2          1       1          0       Inf 0.2727478
#> 11:      2           2          2       1          0       Inf 0.2625399
#> 12:      2           2          3       1          0       Inf 0.3325059
#>         s1.s2 s2.s1 s2.s2
#>  1: 0.6080750     0     1
#>  2: 0.6834333     0     1
#>  3: 0.6600761     0     1
#>  4: 0.6657208     0     1
#>  5: 0.6507399     0     1
#>  6: 0.6966393     0     1
#>  7: 0.6497777     0     1
#>  8: 0.6289767     0     1
#>  9: 0.6732882     0     1
#> 10: 0.7272522     0     1
#> 11: 0.7374601     0     1
#> 12: 0.6674941     0     1