Convert multi-state data to PFS and OS data

Convert a multi-state dataset with irreversible transitions containing 3 health states to a dataset with one row per patient and progression-free survival (PFS) and overall survival (OS) time-to-event outcomes.

as_pfs_os(
  data,
  patient_vars,
  status = "status",
  time_stop = "time_stop",
  transition = "transition_id"
)

Arguments

data
A multi-state dataset.

patient_vars
Character vector of the names of patient specific variables.

status
Character string with the name of the status variable (1 = event, 0 = censored).

time_stop
Character string with the name of the stopping time variable (i.e., time patient transitions from state \(r\) to state \(s\)).

transition
Character string with the name of the variable identifying a transition. The transition variable should be integer valued with values 1, 2, and 3 for the Stable -> Progression, Stable -> Death, and Progression -> Death transitions, respectively.

Value

A data.table with one row per patient containing each variable in patient_vars as well as a time variable and status indicator for both PFS (pfs_status, pfs_time) and OS (os_time, os_status).

Examples

as_pfs_os(onc3, patient_vars = c("patient_id", "female", "age"))

#>       patient_id female      age   pfs_time pfs_status os_status    os_time
#>    1:          1      0 66.40410  9.2091431          1         1  9.2091431
#>    2:          2      0 57.64656  8.5141360          1         1 14.0516883
#>    3:          3      1 59.84345  7.3551054          0         0  7.3551054
#>    4:          4      1 67.21104  2.2942091          0         0  2.2942091
#>    5:          5      0 62.60716 12.2578886          1         1 12.2578886
#>   ---                                                                      
#> 1996:       1996      0 73.19935 15.0000000          0         0 15.0000000
#> 1997:       1997      1 55.03264 11.5105675          1         0 15.0000000
#> 1998:       1998      1 52.87658  0.4990468          0         0  0.4990468
#> 1999:       1999      0 56.87138  7.1510598          1         0 15.0000000
#> 2000:       2000      1 60.93386  6.0576762          0         0  6.0576762