Create CohortDtstmTrans object

A generic function for creating an object of class CohortDtstmTrans.

create_CohortDtstmTrans(object, ...)

# S3 method for multinom_list
create_CohortDtstmTrans(
  object,
  input_data,
  trans_mat,
  n = 1000,
  point_estimate = FALSE,
  ...
)

Arguments

object
An object of the appropriate class.

...
Further arguments passed to CohortDtstmTrans$new() in CohortDtstmTrans.

input_data
An object of class expanded_hesim_data returned by expand.hesim_data()

trans_mat
A transition matrix describing the states and transitions in a discrete-time multi-state model. See CohortDtstmTrans.

n
Number of random observations of the parameters to draw.

point_estimate
If TRUE, then the point estimates are returned and and no samples are drawn.