Create a parameter object from a fitted model

create_params is a generic function for creating an object containing parameters from a fitted statistical model. If uncertainty != "none", then random samples from suitable probability distributions are returned.

create_params(object, ...)

# S3 method for class 'lm'
create_params(object, n = 1000, uncertainty = c("normal", "none"), ...)

# S3 method for class 'multinom'
create_params(object, n = 1000, uncertainty = c("normal", "none"), ...)

# S3 method for class 'multinom_list'
create_params(object, n = 1000, uncertainty = c("normal", "none"), ...)

# S3 method for class 'flexsurvreg'
create_params(object, n = 1000, uncertainty = c("normal", "none"), ...)

# S3 method for class 'flexsurvreg_list'
create_params(object, n = 1000, uncertainty = c("normal", "none"), ...)

# S3 method for class 'partsurvfit'
create_params(
  object,
  n = 1000,
  uncertainty = c("normal", "bootstrap", "none"),
  max_errors = 0,
  silent = FALSE,
  ...
)

Arguments

object

A statistical model to randomly sample parameters from.

...

Currently unused.

n

Number of random observations to draw. Not used if uncertainty = "none".

uncertainty

Method determining how parameter uncertainty should be handled. If "normal", then parameters are randomly drawn from their multivariate normal distribution. If "bootstrap", then parameters are bootstrapped using bootstrap. If "none", then only point estimates are returned.

max_errors

Maximum number of errors that are allowed when fitting statistical models during the bootstrap procedure. This argument may be useful if, for instance, the model fails to converge during some bootstrap replications. Default is 0.

silent

Logical indicating whether error messages should be suppressed. Passed to the silent argument of try().

Value

An object prefixed by params_. Mapping between create_params and the classes of the returned objects are:

create_params.lm -> params_lm

create_params.multinom -> params_mlogit

create_params.multinom_list -> params_mlogit_list

create_params.flexsurvreg -> params_surv

create_params.flexsurvreg_list -> params_surv_list

create_params.partsurvfit -> params_surv_list

See also

Examples

# create_params.lm
fit <- lm(costs ~ female, data = psm4_exdata$costs$medical)
n <- 5
params_lm <- create_params(fit, n = n)
head(params_lm$coefs)
#>      (Intercept)    female
#> [1,]    28628.35 7038.5227
#> [2,]    30309.50 3072.4469
#> [3,]    31059.05 4468.9643
#> [4,]    32941.25  463.5608
#> [5,]    31121.34 3716.2031
head(params_lm$sigma)
#> [1] 10311.3 10311.3 10311.3 10311.3 10311.3

# create_params.flexsurvreg
library("flexsurv")
fit <- flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "weibull")
n <- 5
params_surv_wei <- create_params(fit, n = n)
print(params_surv_wei$dist)
#> [1] "weibull.quiet"
head(params_surv_wei$coefs)
#> $shape
#>      (Intercept)
#> [1,]   0.1181375
#> [2,]   0.3542433
#> [3,]  -0.0513738
#> [4,]  -0.2275955
#> [5,]  -0.1816943
#> 
#> $scale
#>      (Intercept)
#> [1,]    7.394033
#> [2,]    7.243726
#> [3,]    7.302485
#> [4,]    7.299298
#> [5,]    6.902651
#>