Define and evaluate model expression

A model expression is defined by specifying random number generation functions for a probabilistic sensitivity analysis and transformations of the sampled parameters as a function of input_data. The unevaluated expressions are evaluated with eval_model() and used to generate the model inputs needed to create an economic model.

define_model(tparams_def, rng_def, params = NULL, n_states = NULL)

eval_model(x, input_data)

Arguments

tparams_def
A tparams_def object or a list of tparams_def objects. A list might be considered if time intervals specified with the times argument in define_tparams() vary across parameters. Parameters for a transition probability matrix (tpmatrix), utilities (utility), and/or cost categories (costs) are returned as a named list (see define_tparams() for more details).

rng_def
A rng_def object used to randomly draw samples of the parameters from suitable probability distributions.

params
A list containing the values of parameters for random number generation.

n_states
The number of health states (inclusive of all health states including the the death state) in the model. If tpmatrix is an element returned by tparams_def, then it will be equal to the number of states in the transition probability matrix; otherwise it must be specified as an argument.

x
An object of class model_def created with define_model().

input_data
An object of class expanded_hesim_data expaned by patients and treatment strategies.

Value

define_model() returns an object of class model_def, which is a list containing the arguments to the function. eval_model() returns a list containing ID variables identifying parameter samples, treatment strategies, patient cohorts, and time intervals; the values of parameters of the transition probability matrix, utilities, and/or cost categories; the number of health states; and the number of random number generation samples for the PSA.

Details

eval_model() evaluates the expressions in an object of class model_def returned by define_model() and is, in turn, used within functions that instantiate economic models (e.g., create_CohortDtstm()). The direct output of eval_model() can also be useful for understanding and debugging model definitions, but it is not used directly for simulation.

A model_def object would typically be defined in four steps by specifying:

Data (input_data) of class expanded_hesim_data consisting of the treatment strategies and patient population

Parameters (params) objects storing the values of all parameters used in the model

Random number generation (define_rng()) expressions that randomly sample values of the parameters from suitable probability distributions for probabilistic sensitivity analysis

Transformed parameter (define_tparams()) expressions that transform the parameter estimates into values used for simulation

Step 2 can be omitted if underlying parameter values are defined in steps 3. The output of step 4 is used to instantiate the economic model (or a portion of an economic model) as a function of a transition probability matrix, utilities, and/or costs.

See also

Examples

# Data
library("data.table")
strategies <- data.table(strategy_id = 1:2,
                         strategy_name = c("Monotherapy", "Combination therapy"))
patients <- data.table(patient_id = 1)
hesim_dat <- hesim_data(strategies = strategies,
                       patients = patients)
data <- expand(hesim_dat)

# Define the model
rng_def <- define_rng({
  alpha <- matrix(c(1251, 350, 116, 17,
                    0, 731, 512, 15,
                    0, 0, 1312, 437,
                    0, 0, 0, 469),
                  nrow = 4, byrow = TRUE)
  rownames(alpha) <- colnames(alpha) <- c("A", "B", "C", "D")
  lrr_mean <- log(.509)
  lrr_se <- (log(.710) - log(.365))/(2 * qnorm(.975))
  list(
    p_mono = dirichlet_rng(alpha),
    rr_comb = lognormal_rng(lrr_mean, lrr_se),
    u = 1,
    c_zido = 2278,
    c_lam = 2086.50,
    c_med = gamma_rng(mean = c(A = 2756, B = 3052, C = 9007),
                        sd = c(A = 2756, B = 3052, C = 9007))
  )
}, n = 2)

tparams_def <- define_tparams({
  rr = ifelse(strategy_name == "Monotherapy", 1, rr_comb)
  list(
    tpmatrix = tpmatrix(
      C, p_mono$A_B * rr, p_mono$A_C * rr, p_mono$A_D * rr,
      0, C, p_mono$B_C * rr, p_mono$B_D * rr,
      0, 0, C, p_mono$C_D * rr,
      0, 0, 0, 1),
    utility = u,
    costs = list(
      drug = ifelse(strategy_name == "Monotherapy",
                    c_zido, c_zido + c_lam),
      medical = c_med
    )
  )
})

model_def <- define_model(
  tparams_def = tparams_def,
  rng_def = rng_def)

# Evaluate the model expression to generate model inputs
# This can be useful for understanding the output of a model expression
eval_model(model_def, data)

#> $id
#> $id[[1]]
#>    sample strategy_id patient_id time_start
#> 1:      1           1          1          0
#> 2:      1           2          1          0
#> 3:      2           1          1          0
#> 4:      2           2          1          0
#> 
#> 
#> $tpmatrix
#>        s1_s1      s1_s2      s1_s3       s1_s4 s2_s1     s2_s2     s2_s3
#> 1: 0.7268101 0.19442450 0.06471782 0.014047611     0 0.6028085 0.3911359
#> 2: 0.8906267 0.07783907 0.02591018 0.005624049     0 0.8409819 0.1565937
#> 3: 0.7351642 0.18437772 0.06866600 0.011792080     0 0.5873485 0.4016351
#> 4: 0.8745787 0.08731786 0.03251894 0.005584510     0 0.8045759 0.1902069
#>          s2_s4 s3_s1 s3_s2     s3_s3     s3_s4 s4_s1 s4_s2 s4_s3 s4_s4
#> 1: 0.006055519     0     0 0.7496222 0.2503778     0     0     0     1
#> 2: 0.002424365     0     0 0.8997597 0.1002403     0     0     0     1
#> 3: 0.011016407     0     0 0.7771731 0.2228269     0     0     0     1
#> 4: 0.005217166     0     0 0.8944733 0.1055267     0     0     0     1
#> 
#> $utility
#> [1] 1 1 1 1
#> attr(,"id_index")
#> [1] 1
#> 
#> $costs
#> $costs$drug
#> [1] 2278.0 4364.5 2278.0 4364.5
#> attr(,"id_index")
#> [1] 1
#> 
#> $costs$medical
#>            A         B        C
#> 1: 2026.8388  3160.886 10739.12
#> 2: 2026.8388  3160.886 10739.12
#> 3:  702.0553 11787.506 20889.12
#> 4:  702.0553 11787.506 20889.12
#> 
#> 
#> $n_states
#> [1] 4
#> 
#> $n
#> [1] 2
#> 
#> attr(,"class")
#> [1] "eval_model"

# Create an economic model with a factory function
econmod <- create_CohortDtstm(model_def, data)