Random generation for generalized gamma distribution

Draw random samples from a generalized gamma distribution using the parameterization from flexsurv. Written in C++ for speed. Equivalent to flexsurv::rgengamma.

fast_rgengamma(n, mu = 0, sigma = 1, Q)

Arguments

n
Number of random observations to draw.

mu
Vector of location parameters. and columns correspond to rates during specified time intervals.

sigma
Vector of scale parameters as described in flexsurv.

Q
Vector of shape parameters.

Value

A vector of random samples from the generalized gamma distribution. The length of the sample is determined by n. The numerical arguments other than n are recycled so that the number of samples is equal to n.

Examples

n <- 1000
m <- 2 ; s <- 1.7; q <- 1
ptm <- proc.time()
r1 <- fast_rgengamma(n, mu = m, sigma = s, Q = q)
proc.time() - ptm

#>    user  system elapsed 
#>   0.001   0.000   0.001 

ptm <- proc.time()
library("flexsurv")
r2 <- flexsurv::rgengamma(n, mu = m, sigma = s, Q = q)
proc.time() - ptm

#>    user  system elapsed 
#>   0.003   0.000   0.003 

summary(r1)

#>      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
#>   0.00001   0.86459   3.77216  10.93777  12.50414 183.19929 

summary(r2)

#>      Min.   1st Qu.    Median      Mean   3rd Qu.      Max. 
#>   0.00007   0.69912   3.12032  10.28442  11.67629 183.13027