ICER table

Generate a table of incremental cost-effectiveness ratios given output from icea_pw.

icer_tbl(
  x,
  k = 50000,
  cri = TRUE,
  prob = 0.95,
  digits_qalys = 2,
  digits_costs = 0,
  output = c("matrix", "data.table"),
  rownames = NULL,
  colnames = NULL,
  drop = TRUE
)

Arguments

x
An object of class "icea_pw" returned by icea_pw.

k
Willingness to pay.

cri
If TRUE, credible intervals are computed; otherwise they are not.

prob
A numeric scalar in the interval (0,1) giving the credible interval. Default is 0.95 for a 95 percent credible interval.

digits_qalys
Number of digits to use to report QALYs.

digits_costs
Number of digits to use to report costs.

output
Should output be a data.table or a list of matrices for each group.

rownames
Row names for matrices when output = "matrix".

colnames
Column names for matrices when output = "matrix".

drop
If TRUE, then the result is coerced to the lowest possible dimension. Relevant if output = "matrix" and there is one group, in which case a single matrix will be returned if drop = TRUE and a list of length 1 will be returned if drop = FALSE.

Value

If output = "matrix", then a list of matrices (or a matrix if drop = TRUE) reporting incremental cost-effectiveness ratios (ICERs) by group. Specifically, each matrix contains five rows for: (i) incremental quality-adjusted life-years (QALYs), (ii) incremental costs, (iii) the incremental net monetary benefit (NMB), (iv) the ICER, and (v) a conclusion stating whether each strategy is cost-effective relative to a comparator. The number of columns is equal to the number of strategies (including the comparator).

If output = "data.table", then the results are reported as a data.table, with one row for each strategy and group combination.

See also