Incremental treatment effect

Calculate incremental effect of all treatment strategies on outcome variables from probabilistic sensitivity analysis relative to comparator.

incr_effect(x, comparator, sample, strategy, grp, outcomes)

Arguments

x
A data.frame or data.table containing simulation output with information on outcome variables for each randomly sampled parameter set from a PSA. Each row should denote a randomly sampled parameter set and treatment strategy.

comparator
Name of comparator strategy.

sample
Name of column denoting a randomly sampled parameter set.

strategy
Name of column denoting treatment strategy.

grp
Name of column denoting subgroup.

outcomes
Name of columns to calculate incremental changes for.

Value

A data.table containing the differences in the values of each variable specified in outcomes between each treatment strategy and the comparator. It is the same output from delta generated by icea_pw.

Examples

# simulation output
n_samples <- 100
sim <- data.frame(sample = rep(seq(n_samples), 4),
             c = c(rlnorm(n_samples, 5, .1), rlnorm(n_samples, 5, .1),
                    rlnorm(n_samples, 11, .1), rlnorm(n_samples, 11, .1)),
             e = c(rnorm(n_samples, 8, .2), rnorm(n_samples, 8.5, .1),
                   rnorm(n_samples, 11, .6), rnorm(n_samples, 11.5, .6)),
             strategy = rep(paste0("Strategy ", seq(1, 2)),
                           each = n_samples * 2),
             grp = rep(rep(c("Group 1", "Group 2"),
                           each = n_samples), 2)
)
# calculate incremental effect of Strategy 2 relative to Strategy 1 by group
ie <- incr_effect(sim, comparator = "Strategy 1", sample = "sample",
                        strategy = "strategy", grp = "grp", outcomes = c("c", "e"))
head(ie)

#>    sample   strategy     grp       ic       ie
#> 1:      1 Strategy 2 Group 1 66900.32 3.480183
#> 2:      1 Strategy 2 Group 2 56040.03 3.016152
#> 3:      2 Strategy 2 Group 1 60835.65 3.223087
#> 4:      2 Strategy 2 Group 2 58749.91 3.497070
#> 5:      3 Strategy 2 Group 1 65314.95 3.614489
#> 6:      3 Strategy 2 Group 2 65492.68 3.446843