Incremental treatment effect

Calculate incremental effect of all treatment strategies on outcome variables from probabilistic sensitivity analysis relative to comparator.

incr_effect(x, comparator, sim, strategy, grp, outcomes)

Arguments

x
Matrix containing information on outcome variables for each simulation and strategy.

comparator
Name of comparator strategy.

sim
Name of column denoting simulation number.

strategy
Name of column denoting treatment strategy.

grp
Name of column denoting subgroup.

outcomes
Name of columns to calculate incremental changes for.

Value

A data.table containing the differences in the values of each variable specified in outcomes between each treatment strategy and the comparator. It is the same output generated in delta from icea_pw.

Examples

# simulation output
nsims <- 100
sim <- data.frame(sim = rep(seq(nsims), 4),
             c = c(rlnorm(nsims, 5, .1), rlnorm(nsims, 5, .1),
                    rlnorm(nsims, 11, .1), rlnorm(nsims, 11, .1)),
             e = c(rnorm(nsims, 8, .2), rnorm(nsims, 8.5, .1),
                   rnorm(nsims, 11, .6), rnorm(nsims, 11.5, .6)),
             strategy = rep(paste0("Strategy ", seq(1, 2)),
                           each = nsims * 2),
             grp = rep(rep(c("Group 1", "Group 2"),
                           each = nsims), 2)
)
# calculate incremental effect of Strategy 2 relative to Strategy 1 by group
incr.effect <- incr_effect(sim, comparator = "Strategy 1", sim = "sim",
                        strategy = "strategy", grp = "grp", outcomes = c("c", "e"))
head(incr.effect)

#>    sim   strategy     grp       ic       ie
#> 1:   1 Strategy 2 Group 1 52507.73 3.331170
#> 2:   1 Strategy 2 Group 2 46899.64 2.132974
#> 3:   2 Strategy 2 Group 1 47283.51 3.146921
#> 4:   2 Strategy 2 Group 2 58043.04 3.941819
#> 5:   3 Strategy 2 Group 1 60275.34 3.206888
#> 6:   3 Strategy 2 Group 2 66920.52 4.079993