Parameters of a linear model

Create a list containing the parameters of a fitted linear regression model.

params_lm(coefs, sigma = 1)

Arguments

coefs

Samples of the coefficients under sampling uncertainty. Must be a matrix or any object coercible to a matrix such as data.frame or data.table.

sigma

A vector of samples of the standard error of the regression model. Default value is 1 for all samples. Only used if the model is used to randomly simulate values (rather than to predict means).

Value

An object of class params_lm, which is a list containing coefs, sigma, and n_samples. n_samples is equal to the number of rows in coefs. The coefs element is always converted into a matrix.

Details

Fitted linear models are used to predict values, \(y\), as a function of covariates, \(x\), $$y = x^T\beta + \epsilon.$$ Predicted means are given by \(x^T\hat{\beta}\) where \(\hat{\beta}\) is the vector of estimated regression coefficients. Random samples are obtained by sampling the error term from a normal distribution, \(\epsilon \sim N(0, \hat{\sigma}^2)\).

See also

Examples

library("MASS")
n <- 2
params <- params_lm(
  coefs = mvrnorm(n, mu = c(.5,.6),
                  Sigma = matrix(c(.05, .01, .01, .05), nrow = 2)),
  sigma <- rgamma(n, shape = .5, rate = 4)
)
summary(params)
#>    parameter   term      mean        sd       2.5%     97.5%
#>       <char> <char>     <num>     <num>      <num>     <num>
#> 1:      mean     x1 0.3969149 0.3114263 0.18771386 0.6061160
#> 2:      mean     x2 0.5852242 0.2348049 0.42749371 0.7429547
#> 3:        sd  sigma 0.4493351 0.5921983 0.05152505 0.8471452
params
#> A "params_lm" object
#> 
#> Summary of coefficients:
#>    parameter   term      mean        sd      2.5%     97.5%
#>       <char> <char>     <num>     <num>     <num>     <num>
#> 1:      mean     x1 0.3969149 0.3114263 0.1877139 0.6061160
#> 2:      mean     x2 0.5852242 0.2348049 0.4274937 0.7429547
#> 
#> Summary of sigma:
#>    parameter   term      mean        sd       2.5%     97.5%
#>       <char> <char>     <num>     <num>      <num>     <num>
#> 1:        sd  sigma 0.4493351 0.5921983 0.05152505 0.8471452