Parameters of a survival model

Create a list containing the parameters of a single fitted parametric survival model.

params_surv(coefs, dist, aux = NULL)

Arguments

coefs
A list of length equal to the number of parameters in the survival distribution. Each element of the list is a matrix of samples from the posterior distribution of the regression coefficients used to predict a given parameter.

dist
Character vector denoting the parametric distribution. See "Details".

aux
Auxillary arguments used with splines or fractional polynomials. See "Details".

Value

An object of class "params_surv", which is a list containing coefs, dist, and n_samples. n_samples is equal to the number of rows in each element of coefs, which must be the same. The list may also contain aux if a spline or fractional polynomial model is fit.

Details

The types of distributions that can be specified are:

exponential or exp Exponential distribution. coef must contain the rate parameter on the log scale and the same parameterization as in Exponential.

weibull or weibull.quiet Weibull distribution. The first element of coef is the shape parameter (on the log scale) and the second element is the scale parameter (also on the log scale). The parameterization is that same as in Weibull.

gamma Gamma distribution. The first element of coef is the shape parameter (on the log scale) and the second element is the rate parameter (also on the log scale). The parameterization is that same as in GammaDist.

lnorm Lognormal distribution. The first element of coef is the meanlog parameter (i.e., the mean on the log scale) and the second element is the sdlog parameter (i.e., the standard deviation on the log scale). The parameterization is that same as in Lognormal.

gompertz Gompertz distribution. The first element of coef is the shape parameter and the second element is the rate parameter (on the log scale). The parameterization is that same as in Gompertz.

llogis Log-logistic distribution. The first element of coef is the shape parameter (on the log scale) and the second element is the scale parameter (also on the log scale). The parameterization is that same as in Llogis.

gengamma Generalized gamma distribution. The first element of coef is the location parameter mu, the second element is the scale parameter sigma (on the log scale), and the third element is the shape parameter Q. The parameterization is that same as in GenGamma.

survspline Survival splines. Each element of coef is a parameter of the spline model (i.e. gamma_0, gamma_1, \(\ldots\)) with length equal to the number of knots (including the boundary knots). See below for details on the auxillary arguments. The parameterization is that same as in Survspline.

fracpoly Fractional polynomials. Each element of coef is a parameter of the fractional polynomial model (i.e. gamma_0, gamma_1, \(\ldots\)) with length equal to the number of powers minus 1. See below for details on the auxillary arguments (i.e., powers).

Auxillary arguments for spline models should be specified as a list containing the elements:

knots

A numeric vector of knots.

scale

The survival outcome to be modeled as a spline function. Options are "log_cumhazard" for the log cummulative hazard; "log_hazard" for the log hazard rate; "log_cumodds" for the log cummulative odds; and "inv_normal" for the inverse normal distribution function.

timescale

If "log" (the default), then survival is modeled as a spline function of log time; if "identity", then it is modeled as a spline function of time.

Auxillary arguments for fractional polynomial models should be specified as a list containing the elements:

powers

A vector of the powers of the fractional polynomial with each element chosen from the following set: -2. -1, -0.5, 0, 0.5, 1, 2, 3.

Furthermore, when either splines or fractional polynomials are used, the following additional auxillary arguments can be specified:

cumhaz_method

Numerical method used to compute cumulative hazard (i.e., to integrate the hazard function). Options are "quad" for adaptive quadrature and "riemann" for Riemann sum.

step

Step size for computation of cumulative hazard with numerical integration. Only required when integrating using "riemann". No step size is required for "quad".

random_method

Method used to randomly draw from an arbitrary survival function. Options are "invcdf" for the inverse CDF and "sample" for randomly sampling from discrete survival probabilities. When "sample" is chosen, the cumulative hazard must be integrated using "riemann".

Examples

library("flexsurv")
fit <- flexsurvreg(Surv(futime, fustat) ~ 1, data = ovarian, dist = "weibull")
params <- params_surv(coefs = list(shape = fit$res.t["shape", "est", drop = FALSE],
                                   scale = fit$res.t["scale", "est", drop = FALSE]),
                     dist = fit$dlist$name)
print(params)

#> $coefs
#> $coefs$shape
#>             est
#> shape 0.1026105
#> 
#> $coefs$scale
#>            est
#> scale 7.111038
#> 
#> 
#> $dist
#> [1] "weibull.quiet"
#> 
#> $n_samples
#> [1] 1
#> 
#> attr(,"class")
#> [1] "params_surv"