Example data for a 4-state partitioned survival model

A collection of example datasets containing simulated survival, costs, and utility data for a 4-state partitioned survival model.

psm4_exdata

Format

A list containing the following elements:

Survival A data frame containing patient information and time to 3 separate survival endpoints.

CostsA list of data frames. The first data frame contains medical cost data and the second data frame contains drug cost data.

Survival data

The survival data frame contains a list of 3 survival curves, each containing the following columns.

female

An indicator variable equal to 1 if the patient is female and 0 otherwise.

age

The age of the patient in years.

strategy_id

The id of the treatment strategy used.

endpoint1_time

Follow up time with right censored data to survival endpoint 1.

endpoint1_status

A status indicator for survival endpoint 1 equal to 0 if alive and 1 if dead.

endpoint2_time

Follow up time with right censored data to survival endpoint 2.

endpoint2_status

A status indicator for survival endpoint 2 equal to 0 if alive and 1 if dead.

endpoint3_time

Follow up time with right censored data to survival endpoint 3.

endpoint3_status

A status indicator for survival endpoint 3 equal to 0 if alive and 1 if dead.

Cost data

The cost list contains two data frames.The first data frame contains data on the medical costs by patient and health state, and contains the following columns:

patient_id

An integer denoting the id of the patient.

female

An indicator variable equal to 1 if the patient is female and 0 otherwise.

state_name

A categorical variable denoting the three possible health states.

costs

Annualized medical costs.

The second data frame contains data on the drug costs associated with each treatment strategy.

strategy_id

The id of each treatment strategy.

costs

Annualized drug costs.