Summarize transition probability matrix

Summarize a tpmatrix object storing transition probability matrices. Summary statistics are computed for each possible transition.

# S3 method for tpmatrix
summary(object, id = NULL, probs = c(0.025, 0.975), unflatten = FALSE, ...)

Arguments

object
A tpmatrix object.

id
A tpmatrix_id object for which columns contain the ID variables for each row in object. If not NULL, then transition probability matrices are summarized by the ID variables in id.

probs
A numeric vector of probabilities with values in [0,1] used to compute quantiles. By default, the 2.5th and 97.5th percentiles are computed.

unflatten
If FALSE, then each column containing a summary statistic is a vector and the generated table contains one row (for each set of ID variables) for each possible transition; if TRUE, then each column stores a list of matrix objects containing transition probability matrices formed by "unflattening" the one-dimensional vectors. See "Value" below for additional details.

...
Additional arguments affecting the summary. Currently unused.

Value

If unflatten = "FALSE" (the default), then a data.table is returned with columns for (i) the name of the parameter representing a transition in the transition probability matrix (params), (ii) the mean of each parameter across parameter samples (mean), (iii) the standard deviation of the parameter samples (sd), and (iv) quantiles of the parameter samples corresponding to the probs argument.

If, on the other hand, unflatten = "TRUE", then the parameters are unflattened to form transition probability matrices; that is, the mean, sd, and quantile columns are matrices.

Examples

library("data.table")
hesim_dat <-  hesim_data(strategies = data.table(strategy_id = 1:2),
                                patients = data.table(patient_id = 1:3))
input_data <- expand(hesim_dat, by = c("strategies", "patients"))

# Summarize across all rows in "input_data"
p_12 <- ifelse(input_data$strategy_id == 1, .8, .6)
p <- tpmatrix(
  C, p_12,
  0, 1
)

## Summary where each column is a vector
summary(p)

#>    param mean        sd 2.5% 97.5%
#> 1: s1_s1  0.3 0.1095445  0.2   0.4
#> 2: s1_s2  0.7 0.1095445  0.6   0.8
#> 3: s2_s1  0.0 0.0000000  0.0   0.0
#> 4: s2_s2  1.0 0.0000000  1.0   1.0

## Summary where each column is a matrix
ps <- summary(p, probs = .5, unflatten = TRUE)
ps

#>               mean                                      sd             50%
#> 1: 0.3,0.0,0.7,1.0 0.1095445,0.0000000,0.1095445,0.0000000 0.3,0.0,0.7,1.0

ps$mean

#> [[1]]
#>      [,1] [,2]
#> [1,]  0.3  0.7
#> [2,]  0.0  1.0
#> 

# Summarize by ID variables
tpmat_id <- tpmatrix_id(input_data, n_samples = 2) 
p_12 <- ifelse(tpmat_id$strategy_id == 1, .8, .6)
p <- tpmatrix(
  C, p_12,
  0, 1
)

## Summary where each column is a vector
summary(p, id = tpmat_id)

#>     strategy_id patient_id param mean sd 2.5% 97.5%
#>  1:           1          1 s1_s1  0.2  0  0.2   0.2
#>  2:           1          1 s1_s2  0.8  0  0.8   0.8
#>  3:           1          1 s2_s1  0.0  0  0.0   0.0
#>  4:           1          1 s2_s2  1.0  0  1.0   1.0
#>  5:           1          2 s1_s1  0.2  0  0.2   0.2
#>  6:           1          2 s1_s2  0.8  0  0.8   0.8
#>  7:           1          2 s2_s1  0.0  0  0.0   0.0
#>  8:           1          2 s2_s2  1.0  0  1.0   1.0
#>  9:           1          3 s1_s1  0.2  0  0.2   0.2
#> 10:           1          3 s1_s2  0.8  0  0.8   0.8
#> 11:           1          3 s2_s1  0.0  0  0.0   0.0
#> 12:           1          3 s2_s2  1.0  0  1.0   1.0
#> 13:           2          1 s1_s1  0.4  0  0.4   0.4
#> 14:           2          1 s1_s2  0.6  0  0.6   0.6
#> 15:           2          1 s2_s1  0.0  0  0.0   0.0
#> 16:           2          1 s2_s2  1.0  0  1.0   1.0
#> 17:           2          2 s1_s1  0.4  0  0.4   0.4
#> 18:           2          2 s1_s2  0.6  0  0.6   0.6
#> 19:           2          2 s2_s1  0.0  0  0.0   0.0
#> 20:           2          2 s2_s2  1.0  0  1.0   1.0
#> 21:           2          3 s1_s1  0.4  0  0.4   0.4
#> 22:           2          3 s1_s2  0.6  0  0.6   0.6
#> 23:           2          3 s2_s1  0.0  0  0.0   0.0
#> 24:           2          3 s2_s2  1.0  0  1.0   1.0
#>     strategy_id patient_id param mean sd 2.5% 97.5%

## Summary where each column is a matrix
ps <- summary(p, id = tpmat_id, unflatten = TRUE)
ps

#>    strategy_id patient_id            mean      sd            2.5%
#> 1:           1          1 0.2,0.0,0.8,1.0 0,0,0,0 0.2,0.0,0.8,1.0
#> 2:           1          2 0.2,0.0,0.8,1.0 0,0,0,0 0.2,0.0,0.8,1.0
#> 3:           1          3 0.2,0.0,0.8,1.0 0,0,0,0 0.2,0.0,0.8,1.0
#> 4:           2          1 0.4,0.0,0.6,1.0 0,0,0,0 0.4,0.0,0.6,1.0
#> 5:           2          2 0.4,0.0,0.6,1.0 0,0,0,0 0.4,0.0,0.6,1.0
#> 6:           2          3 0.4,0.0,0.6,1.0 0,0,0,0 0.4,0.0,0.6,1.0
#>              97.5%
#> 1: 0.2,0.0,0.8,1.0
#> 2: 0.2,0.0,0.8,1.0
#> 3: 0.2,0.0,0.8,1.0
#> 4: 0.4,0.0,0.6,1.0
#> 5: 0.4,0.0,0.6,1.0
#> 6: 0.4,0.0,0.6,1.0

ps$mean

#> [[1]]
#>      [,1] [,2]
#> [1,]  0.2  0.8
#> [2,]  0.0  1.0
#> 
#> [[2]]
#>      [,1] [,2]
#> [1,]  0.2  0.8
#> [2,]  0.0  1.0
#> 
#> [[3]]
#>      [,1] [,2]
#> [1,]  0.2  0.8
#> [2,]  0.0  1.0
#> 
#> [[4]]
#>      [,1] [,2]
#> [1,]  0.4  0.6
#> [2,]  0.0  1.0
#> 
#> [[5]]
#>      [,1] [,2]
#> [1,]  0.4  0.6
#> [2,]  0.0  1.0
#> 
#> [[6]]
#>      [,1] [,2]
#> [1,]  0.4  0.6
#> [2,]  0.0  1.0
#>