Predicted means

Create a list containing means predicted from a statistical model.

tparams_mean(value, ...)

Arguments

value
Matrix of samples from the distribution of the mean. Columns denote random samples and rows denote means for different observations.

...
Arguments to pass to id_attributes. Each row in value must be a prediction for a strategy_id, patient_id, state_id, and optionally time_id combination.

Value

An object of class tparams_mean, which is a list containing value, n_samples, and the ID attributes passed to id_attributes.

See also

Examples

tparams_mean(value = matrix(1:8, nrow = 4),
             strategy_id = rep(1:2, each = 2),
             n_strategies = 2,
             patient_id = rep(1, 4),
             n_patients = 1,
             state_id = rep(1:2, times = 2),
             n_states = 2)

#> $value
#>      [,1] [,2]
#> [1,]    1    5
#> [2,]    2    6
#> [3,]    3    7
#> [4,]    4    8
#> 
#> $n_samples
#> [1] 2
#> 
#> $strategy_id
#> [1] 1 1 2 2
#> 
#> $n_strategies
#> [1] 2
#> 
#> $patient_id
#> [1] 1 1 1 1
#> 
#> $n_patients
#> [1] 1
#> 
#> $state_id
#> [1] 1 2 1 2
#> 
#> $n_states
#> [1] 2
#> 
#> $grp_id
#> [1] 1 1 1 1
#> 
#> attr(,"class")
#> [1] "tparams_mean"